% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_required_str}
\alias{add_required_str}
\title{Mark Required Fields with a Prefix}
\usage{
add_required_str(p, r, req_str = "* ")
}
\arguments{
\item{p}{Character vector. All fields to be processed.}

\item{r}{Character vector. Fields that are required.}

\item{req_str}{Character. Prefix to mark required fields. Defaults to \code{"* "}.}
}
\value{
A character vector with required fields prefixed and ordered to
appear before non-required fields.
}
\description{
Mark Required Fields with a Prefix
}
\examples{
# Mark required fields with a prefix
add_required_str(c("Name", "Id", "Age"), c("Id", "Name"))

}
