% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{aliases_formatter}
\alias{aliases_formatter}
\title{Format Aliases from a Table}
\usage{
aliases_formatter(tab, params)
}
\arguments{
\item{tab}{Data frame. The input table where the first row contains column
names.}

\item{params}{List. Additional parameters for formatting. Takes a formatter
params value from parser parameter yaml file. Currently unused.}
}
\value{
A named list where each name corresponds to a formatted column name,
and values are non-NA elements of the respective column.
}
\description{
Format Aliases from a Table
}
\examples{
\dontrun{
# Format aliases from a table
tab <- data.frame(Alias = c("name1", "name2", NA), Value = c(1, 2, 3))
formatted <- aliases_formatter(tab, params = list())
}

}
