% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{api_name_to_label}
\alias{api_name_to_label}
\title{Convert API Names to Prettified Labels}
\usage{
api_name_to_label(x)
}
\arguments{
\item{x}{Character vector. API field names to be converted.}
}
\value{
A character vector with API names converted to human-readable labels.
}
\description{
This function converts API-style names with underscores into human-readable
labels by replacing underscores with spaces and applying title case.
}
\examples{
api_name_to_label(c("first_name", "last_name", "instument_model"))

}
