% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_ega.R
\name{check_list}
\alias{check_list}
\title{Check Properties Of The List}
\usage{
check_list(ll, nms)
}
\arguments{
\item{ll}{List. List to be checked.}

\item{nms}{Character Vector. Required element names in the list.}
}
\value{
List. Checked list with reordered elements according to the  names in \code{nms}.
}
\description{
Helper function to verify names and structure of a list.
}
\examples{
extra <- list(a=1, b=2)
Rega:::check_list(ll=extra, nms=c("a","b"))

}
\keyword{internal}
