% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{.add_paths}
\alias{.add_paths}
\title{Generate URL Parameter Replacement Expressions for an API Request}
\usage{
.add_paths(path_params)
}
\arguments{
\item{path_params}{A character vector of names of the parameters to replace
in the URL. Each name should correspond to a placeholder in the URL in the
format \code{{param}}. Path parameters are created \code{.get_operation_params}
function.}
}
\value{
A list of expressions. Each expression replaces a \code{{param}}
placeholder in the URL with the value of the corresponding parameter.
}
\description{
This function creates a list of expressions to replace placeholders in a URL
with corresponding parameter values.
}
\examples{
# Generate replacement expressions for path parameters
Rega:::.add_paths(c("id", "type"))

}
\keyword{internal}
