% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{.add_request_body}
\alias{.add_request_body}
\title{Add Request Body to API Request}
\usage{
.add_request_body(has_body)
}
\arguments{
\item{has_body}{A logical value indicating whether the request requires a
body.}
}
\value{
An expression to add the JSON request body if \code{has_body} is
\code{TRUE}, otherwise an empty list.
}
\description{
Adds a JSON request body to an API request if required. If the request
requires a body, an expression is returned to include it; otherwise, an empty
list is returned.
}
\examples{
Rega:::.add_request_body(TRUE)
Rega:::.add_request_body(FALSE)

}
\keyword{internal}
