% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.basic_validator}
\alias{.basic_validator}
\title{Create a Basic Validator for Data Columns}
\usage{
.basic_validator(meta, aliases, column, code_list_entry)
}
\arguments{
\item{meta}{A list containing parsed EGA metadata, where elements correspond
sheets, which contain columns to be validated.}

\item{aliases}{A named list of reference values used for validation.}

\item{column}{A string specifying the name of the column to validate.}

\item{code_list_entry}{A string specifying the alias entry in \code{aliases}
and sheet name in metadata to compare against.}
}
\value{
A validation object containing the results of the check or NULL if
sheet specified by \code{ode_list_entry} was not present in metadata
}
\description{
Generates a validator to check multiple conditions on a specified data
column. Following validations are performed:
\itemize{
\item are all entries specified (non-NA)
\item are all entries unique
\item are all entries present in aliases
\item are all aliases present in the sheet
}
}
\examples{
meta <- list(sample_codes = data.frame(sample_ids = c("A", "B", "C")))
aliases <- list(sample_codes = c("A", "B", "C", "D"))
Rega:::.basic_validator(meta, aliases, "sample_ids", "sample_codes")

}
\keyword{internal}
