% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.dataset_analysis_validator}
\alias{.dataset_analysis_validator}
\title{Validate Datasets Analyses Metadata}
\usage{
.dataset_analysis_validator(meta, aliases)
}
\arguments{
\item{meta}{A list containing parsed EGA metadata, where elements correspond
to individual sheets.}

\item{aliases}{A list of aliases, reference values used for validation (not
used in this function but included for consistency with other validators).}
}
\value{
A validation object containing the results of the check or NULL if
analyses sheet was not present in metadata
}
\description{
Validates the metadata by checking the following:
\itemize{
\item analyses for datasets are present in aliases
}
}
\examples{
meta <- list(analyses = data.frame(
   analyses = c("AnalysisA", "AnalysisB", NA)
))

aliases = list(
    analyses = c("AnalysisA", "AnalysisC")
)

Rega:::.dataset_analysis_validator(meta, aliases)

}
\keyword{internal}
