% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\docType{data}
\name{.EGA_TOKEN_URL}
\alias{.EGA_TOKEN_URL}
\title{Default URL for EGA OAuth token}
\format{
An object of class \code{character} of length 1.
}
\usage{
.EGA_TOKEN_URL
}
\description{
Default URL for EGA OAuth token
}
\keyword{internal}
