% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{.get_ega_username}
\alias{.get_ega_username}
\title{Retrieve EGA Username}
\usage{
.get_ega_username(keyring_name = "REGA_EGA", envvar = "REGA_EGA_USERNAME")
}
\arguments{
\item{keyring_name}{Character. Name of the service in the keyring. Defaults
to \code{"REGA_EGA"}.}

\item{envvar}{Character. Name of the environment variable. Defaults to
\code{"REGA_EGA_USERNAME"}.}
}
\value{
A character string containing the username.
}
\description{
Fetches the EGA username from the system keyring, an environment variable, or
an interactive prompt, following that priority order.
}
\examples{
tryCatch(
    Rega:::.get_ega_username("REGA_EGA_USERNAME")
)

}
\keyword{internal}
