% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.has_analyses}
\alias{.has_analyses}
\title{Check for Presence of Analyses Sheet in Metadata}
\usage{
.has_analyses(meta)
}
\arguments{
\item{meta}{A list containing metadata with elements \code{analyses} and
\code{analysis_files}.}
}
\value{
A logical value: \code{TRUE} if analyses are absent or empty,
\code{FALSE} otherwise.
}
\description{
Determines whether analyses or analysis files are specified in the provided
metadata.
}
\examples{
meta <- list(
    aliases = list(analyses = c("A1", "A2")),
    analysis_files = data.frame(name = character(0))
)
Rega:::.has_analyses(meta)

}
\keyword{internal}
