% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_scalar}
\alias{.is_scalar}
\title{Check if an object is a non-complex atomic scalar.}
\usage{
.is_scalar(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A logical value: \code{TRUE} if \code{x} is a non-complex
atomic scalar, \code{FALSE} otherwise.
}
\description{
An atomic scalar is defined as an atomic vector of length exactly
one, excluding complex numbers.
}
\examples{
# Returns TRUE for simple, single values
.is_scalar(5L)
.is_scalar("hello")
.is_scalar(TRUE)
# Returns FALSE for vectors, lists, or complex numbers
.is_scalar(c(1, 2))
.is_scalar(list(1))
.is_scalar(1 + 1i)

}
\keyword{internal}
