% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{.operation_params_to_args}
\alias{.operation_params_to_args}
\title{Convert Operation Parameters to Function Arguments}
\usage{
.operation_params_to_args(op)
}
\arguments{
\item{op}{List. An operation definition containing a \code{parameters} element,
which is a list of parameter definitions. Each parameter should include a
\code{name} and an optional \code{required} field.}
}
\value{
A named list representing function arguments. Names correspond to
parameter names, with required parameters set to missing (\code{quote(expr = )})
}
\description{
This function transforms an operation's parameter definitions into a list of
function arguments. Required parameters are marked as missing arguments.
}
\examples{
# Convert operation parameters to function arguments
opdefs <- extract_operation_definitions(extract_api())

Rega:::.operation_params_to_args(
    opdefs[["post__submissions__provisional_id__samples"]]
)

}
\keyword{internal}
