% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.runs_extra_validator}
\alias{.runs_extra_validator}
\title{Validate Runs Metadata}
\usage{
.runs_extra_validator(meta, aliases)
}
\arguments{
\item{meta}{A list containing parsed EGA metadata, where elements correspond
to individual sheets.}

\item{aliases}{A list of aliases, reference values used for validation.}
}
\value{
A validation object containing the results of the check.
}
\description{
Validates the metadata by checking the following:
\itemize{
\item experiments for runs are specified
\item samples for runs are specified
\item file types for runs are specified
\item files for runs are specified
\item files for runs are unique
\item experiments for runs are present in aliases
\item samples for runs are present in aliases
}
}
\examples{
meta <- list(runs = data.frame(
    experiment = c("ExperimentA", "ExperimentB", NA),
    alias = c("Sample1", "Sample2", "Sample1"),
    run_file_type = c("fastq", NA, "fastq"),
    files = I(list(c("File1", "File2"), c("File3"), NA))
))

aliases <- list(
    experiments = c("ExperimentA", "ExperimentB", "ExperimentC"),
    samples = c("Sample1", "Sample3")
)

validate::summary(Rega:::.runs_extra_validator(meta, aliases))

}
\keyword{internal}
