% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.submission_validator}
\alias{.submission_validator}
\title{Validate Submission Metadata}
\usage{
.submission_validator(meta, aliases)
}
\arguments{
\item{meta}{A list containing parsed EGA metadata, where elements correspond
to individual sheets.}

\item{aliases}{A list of aliases, reference values used for validation (not
used in this function but included for consistency with other validators).}
}
\value{
A validation object containing the results of the check.
}
\description{
Validates the metadata by checking the following:
\itemize{
\item submission title is specified
}
}
\examples{
meta <- list(submission = data.frame(title = c("Submission A", NA)))
Rega:::.submission_validator(meta, list())

}
\keyword{internal}
