% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{.url_param_replacer}
\alias{.url_param_replacer}
\title{Generate URL Parameter Replacement Expressions}
\usage{
.url_param_replacer(path_params, url)
}
\arguments{
\item{path_params}{Character vector. Names of the parameters to replace in
the URL. Each name should correspond to a placeholder in the format
\code{{param}}.}

\item{url}{Character. The URL containing placeholders to be replaced.}
}
\value{
A list of expressions. Each expression replaces a \code{{param}}
placeholder in the URL with the value of the corresponding parameter.
}
\description{
This function creates a list of expressions to replace placeholders in a URL
with corresponding parameter values.
}
\examples{
\dontrun{
# Generate replacement expressions for path parameters
replacers <- Rega:::.url_param_replacer(
  c("id", "type"), "https://api/{id}/{type}"
)
}

}
\keyword{internal}
