% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_logical_scalar}
\alias{.validate_logical_scalar}
\title{Validate that an object is a non-missing logical scalar.}
\usage{
.validate_logical_scalar(x, varname = NULL)
}
\arguments{
\item{x}{An object to be tested.}

\item{varname}{A character string specifying the name of the variable being
checked, used in the error message.  If \code{NULL}, the variable name is
taken from the name of variable being tested. Defaults to \code{NULL}}
}
\value{
Returns \code{TRUE} if the validation succeeds, otherwise stops
execution with an error.
}
\description{
Checks if \code{x} is a logical vector with a length of exactly one and is
not \code{NA}. If the check fails, the function stops execution and reports
an error using the variable name.
}
\examples{
# Returns TRUE silently
Rega:::.validate_logical_scalar(TRUE, "input_check")

# Throws an error for a logical NA
tryCatch(
    Rega:::.validate_logical_scalar(NA, "na_arg"),
    error = function(e) message(e$message)
)

}
\keyword{internal}
