% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_analyses.R
\name{ega_analyses}
\alias{ega_analyses}
\title{List All Analyses in EGA}
\usage{
ega_analyses(submission = NULL, analysis = NULL)
}
\arguments{
\item{submission}{Get information for all analyses in this submission.}

\item{analysis}{Get information only for this analysis.}
}
\value{
tibble with information about analyses.
}
\description{
Get information about existing analyses in EGA. If provisional or permanent accession \code{analysis} id is specified, then only information about this asnalysis is shown.
}
\examples{
\dontrun{
ega_analyses()
}

}
