% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_analyses.R
\name{ega_create_analysis}
\alias{ega_create_analysis}
\title{Create Analysis in EGA}
\usage{
ega_create_analysis(
  submission,
  title,
  description,
  analysis_type,
  files,
  experiment_types = NULL,
  genome_id = NULL,
  chromosomes = NULL,
  platform = NULL,
  study_provisional_id = NULL,
  study_accession_id = NULL,
  experiment_provisional_ids = NULL,
  experiment_accession_ids = NULL,
  sample_provisional_ids = NULL,
  sample_accession_ids = NULL,
  extra_attributes = NULL
)
}
\arguments{
\item{submission}{Character scalar. Provisional or stable submission Id.}

\item{title}{Character scalar. The analysis title.}

\item{description}{Character scalar. The analysis description.}

\item{analysis_type}{Character scalar. The analysis analysis type. See \code{ega_enums("analysis_types")}.}

\item{files}{List. Provisional Ids of the files.}

\item{experiment_types}{Character scalar. Types of experiments.}

\item{genome_id}{Integer scalar. The analysis genome id. See \code{ega_enums("genomes")}.}

\item{chromosomes}{List. See \code{ega_enums("chromosomes")}.}

\item{platform}{Character scalar. The analysis platform.}

\item{study_provisional_id}{Integer scalar. The analysis's study's provisional ID.}

\item{study_accession_id}{Character scalar. The analysis's study's accession ID.}

\item{experiment_provisional_ids}{Integer scalar. The analysis's experiment's provisional ID.}

\item{experiment_accession_ids}{Character scalar. The analysis's experiment's accession ID.}

\item{sample_provisional_ids}{Integer vector. The analysis's sample's accession ID}

\item{sample_accession_ids}{Character scalar. The analysis's sample's accession ID.}

\item{extra_attributes}{List. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}
}
\value{
tibble with information about the created analysis.
}
\description{
Create analysis in EGA by assigning files to samples and experiments.
}
\examples{
\dontrun{
new_analysis <- ega_create_analysis(submission_id=provisional_id,
title="my first analysis", description="description of my first analysis",
files=c("75243", "75406"))
}

}
