% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_datasets.R
\name{ega_create_dataset}
\alias{ega_create_dataset}
\title{Create Dataset in EGA}
\usage{
ega_create_dataset(
  submission,
  title,
  description,
  dataset_types,
  policy_accession_id,
  run_provisional_ids = NULL,
  run_accession_ids = NULL,
  analysis_provisional_ids = NULL,
  analysis_accession_ids = NULL,
  extra_attributes = NULL
)
}
\arguments{
\item{submission}{Character scalar. Provisional or stable submission ID.}

\item{title}{Character scalar. Dataset title.}

\item{description}{Character scalar. Dataset description.}

\item{dataset_types}{Character scalar. The dataset types. See \code{ega_enums("dataset_types")}.}

\item{policy_accession_id}{Character scalar. The policy accession ID.}

\item{run_provisional_ids}{Integer vector. Provisional Run IDs.}

\item{run_accession_ids}{Character vector. Accession Run IDs.}

\item{analysis_provisional_ids}{Integer vector. Provisional Analysis IDs.}

\item{analysis_accession_ids}{Character vector. Accession Analysis IDs.}

\item{extra_attributes}{List. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}
}
\value{
tibble with information about the created dataset.
}
\description{
Create Dataset in EGA by assigning files to samples and experiments.
}
\examples{
\dontrun{
new_dataset <- ega_create_dataset(submission_id=provisional_id,
dataset_file_type="fastq", files=c("75243", "75406"),
experiment_provisional_id=7000, sample_provisional_id=46800)
}

}
