% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_experiments.R
\name{ega_create_experiment}
\alias{ega_create_experiment}
\title{Create experiment In EGA}
\usage{
ega_create_experiment(
  submission,
  study_provisional_id,
  study_accession_id = NULL,
  design_description,
  library_name = NULL,
  library_construction_protocol = NULL,
  paired_nominal_length = NULL,
  paired_nominal_sdev = NULL,
  instrument_model_id,
  library_layout,
  library_strategy,
  library_source,
  library_selection,
  extra_attributes = NULL
)
}
\arguments{
\item{submission}{character scalar. Provisional or stable submission id.}

\item{study_provisional_id}{character scalar. The experiment's study provisional ID. Use this field if the referenced study does not have an accession}

\item{study_accession_id}{character scalar. The experiment's study accession ID. Use this field if the referenced study already has an accession}

\item{design_description}{character scalar. The experiment alias.}

\item{library_name}{character scalar. The experiment title.}

\item{library_construction_protocol}{character scalar.}

\item{paired_nominal_length}{integer scalar. The experiment paired nominal length.}

\item{paired_nominal_sdev}{numeric scalar. The experiment paired nominal sdev.}

\item{instrument_model_id}{character scalar. The experiment instrument model id. See \code{ega_enums("platform_models")}.}

\item{library_layout}{character scalar. The experiment library layout. See \code{ega_enums("library_layouts")}.}

\item{library_strategy}{character scalar. The experiment library strategy. See \code{ega_enums("library_strategies")}.}

\item{library_source}{character scalar. The experiment library source. See \code{ega_enums("library_sources")}.}

\item{library_selection}{character scalar. The experiment library selection. See \code{ega_enums("library_selections")}.}

\item{extra_attributes}{list. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}
}
\value{
\code{tibble} with information about the created experiment.
}
\description{
Create a single experiment in EGA.
}
\examples{
\dontrun{
new_experiment <- ega_create_experiment(submission=provisional_id,
alias="My unique experiment alias 1",
biological_sex="male",
phenotype="nose",
subject_id="192873366738836788")
}
}
