% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_runs.R
\name{ega_create_run}
\alias{ega_create_run}
\title{Create Run in EGA}
\usage{
ega_create_run(
  submission,
  run_file_type,
  files,
  experiment_provisional_id = NULL,
  experiment_accession_id = NULL,
  sample_provisional_id = NULL,
  sample_accession_id = NULL,
  extra_attributes = NULL
)
}
\arguments{
\item{submission}{Character scalar. Provisional or stable submission Id.}

\item{run_file_type}{Character scalar. File type.}

\item{files}{List. Provisional Ids of the files.}

\item{experiment_provisional_id}{Character scalar. Provisional Experiment Id.}

\item{experiment_accession_id}{Character scalar. Experiment Accession Id.}

\item{sample_provisional_id}{Character scalar. Provisional Sample Id.}

\item{sample_accession_id}{character scalar. Sample Accession Id.}

\item{extra_attributes}{List. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}
}
\value{
tibble with information about the created run.
}
\description{
Create Run in EGA by assigning files to samples and experiments.
}
\examples{
\dontrun{
new_run <- ega_create_run(submission_id=provisional_id,
run_file_type="fastq", files=c("75243", "75406"),
experiment_provisional_id=7000, sample_provisional_id=46800)
}

}
