% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_samples.R
\name{ega_create_sample}
\alias{ega_create_sample}
\title{Create Sample In EGA}
\usage{
ega_create_sample(
  submission,
  alias,
  biological_sex,
  subject_id,
  phenotype,
  title = NULL,
  description = NULL,
  biosample_id = NULL,
  case_control = NULL,
  organism_part = NULL,
  cell_line = NULL,
  extra_attributes = NULL
)
}
\arguments{
\item{submission}{character scalar. Provisional or stable submission id.}

\item{alias}{character scalar. Sample alias (must be unique within the study).}

\item{biological_sex}{character scalar. Biological sex of subject, for accepted values check \code{ega_enums("biological_sex")}.}

\item{subject_id}{character scalar. Subject Id.}

\item{phenotype}{character scalar. Phenotype group}

\item{title}{character scalar. Sample title.}

\item{description}{character scalar. Sample description}

\item{biosample_id}{character scalar. Biosample Id.}

\item{case_control}{character scalar. Case or Control, for accepted values check \code{ega_enums("case_controls")}.}

\item{organism_part}{character scalar. Sample tissue or cell type.}

\item{cell_line}{character scalar. Name of teh cell line}

\item{extra_attributes}{list. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}
}
\value{
\code{tibble} with information about the created sample.
}
\description{
Create a single sample in EGA.
}
\section{Functions}{
\itemize{
\item \code{ega_create_sample()}: Create a single sample.

}}
\examples{
\dontrun{
new_sample <- ega_create_sample(submission=provisional_id,
alias="My unique sample alias 1",
biological_sex="male",
phenotype="nose",
subject_id="192873366738836788")
}
}
