% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_studies.R
\name{ega_create_study}
\alias{ega_create_study}
\title{Create A New study In EGA}
\usage{
ega_create_study(
  submission,
  title,
  description,
  study_type,
  pubmed_ids = NULL,
  custom_tags = NULL,
  extra_attributes = NULL,
  repositories = NULL
)
}
\arguments{
\item{submission}{character scalar. Submission Id. Study will be created within this submission.}

\item{title}{character scalar. Study title.}

\item{description}{character scalar. Study description.}

\item{study_type}{Study type, for options check output of \code{ega_enums(enums="study_types")}.}

\item{pubmed_ids}{list. List study PubMedIDs. Single element list without names.}

\item{custom_tags}{list. The study custom tags. Single element list without names.}

\item{extra_attributes}{list. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}

\item{repositories}{list. The external references to other repositories, list with element names: \code{repository_id}, \code{url} and \code{label}.}
}
\value{
tibble. Information about newly created study.
}
\description{
Create A New study In EGA
}
\examples{
\dontrun{
ega_create_study(title="study Title",
description="study description")
}
}
