% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_submissions.R
\name{ega_create_submission}
\alias{ega_create_submission}
\title{Create A New Submission In EGA}
\usage{
ega_create_submission(title, description, collaborators = NULL)
}
\arguments{
\item{title}{character scalar. Submission title.}

\item{description}{character scalar. Submission description.}

\item{collaborators}{list. Default \code{NULL} otherwise \code{list} with three elements: \code{id} containing user id; \code{access_type} granted either "read" or "write"; \code{comment} to collaborator.}
}
\value{
tibble. Information about newly created submission.
}
\description{
Create A New Submission In EGA
}
\examples{
\dontrun{
ega_create_submission(title="Submission Title",
description="Submission description")
}
}
