% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_datasets.R
\name{ega_datasets}
\alias{ega_datasets}
\title{List All Datasets in EGA}
\usage{
ega_datasets(submission = NULL, dataset = NULL)
}
\arguments{
\item{submission}{Get information for all datasets in this submission.}

\item{dataset}{Get information only for this dataset.}
}
\value{
tibble with information about datasets.
}
\description{
Get information about existing datasets in EGA. If provisional or permanent accession \code{dataset} id is specified, then only information about this dataset is shown.
}
\examples{
\dontrun{
ega_datasets()
}

}
