% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_analyses.R
\name{ega_delete_analysis}
\alias{ega_delete_analysis}
\title{Delete Run In EGA}
\usage{
ega_delete_analysis(analysis)
}
\arguments{
\item{analysis}{character scalar. Provisional analysis Id.}
}
\value{
tibble. Information about deleted analysis.
}
\description{
Delete this analysis.
}
\examples{
\dontrun{
ega_delete_analysis(analysis=1)
}
}
