% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_datasets.R
\name{ega_delete_dataset}
\alias{ega_delete_dataset}
\title{Delete Dataset In EGA}
\usage{
ega_delete_dataset(dataset)
}
\arguments{
\item{dataset}{character scalar. Provisional dataset Id.}
}
\value{
tibble. Information about deleted dataset.
}
\description{
Delete this dataset.
}
\examples{
\dontrun{
ega_delete_dataset(dataset=1)
}
}
