% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_experiments.R
\name{ega_delete_experiment}
\alias{ega_delete_experiment}
\title{Delete experiment In EGA}
\usage{
ega_delete_experiment(experiment)
}
\arguments{
\item{experiment}{character scalar. Provisional experiment Id.}
}
\value{
tibble. Information about deleted experiment.
}
\description{
Delete this experiment.
}
\examples{
\dontrun{
ega_delete_experiment(experiment=1)
}
}
