% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_runs.R
\name{ega_delete_run}
\alias{ega_delete_run}
\title{Delete Run In EGA}
\usage{
ega_delete_run(run)
}
\arguments{
\item{run}{character scalar. Provisional run Id.}
}
\value{
tibble. Information about deleted run.
}
\description{
Delete this run.
}
\examples{
\dontrun{
ega_delete_run(run=1)
}
}
