% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_samples.R
\name{ega_delete_sample}
\alias{ega_delete_sample}
\title{Delete Sample In EGA}
\usage{
ega_delete_sample(sample)
}
\arguments{
\item{sample}{character scalar. Provisional sample Id.}
}
\value{
tibble. Information about deleted sample.
}
\description{
Delete this sample.
}
\examples{
\dontrun{
ega_delete_sample(sample=1)
}
}
