% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_studies.R
\name{ega_delete_study}
\alias{ega_delete_study}
\title{Delete study In EGA}
\usage{
ega_delete_study(study)
}
\arguments{
\item{study}{character scalar. Provisional study Id.}
}
\value{
tibble. Information about deleted study.
}
\description{
Delete this study, including all the objects in it.
}
\examples{
\dontrun{
ega_delete_study(study=1)
}
}
