% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_submissions.R
\name{ega_delete_submission}
\alias{ega_delete_submission}
\title{Delete Submission In EGA}
\usage{
ega_delete_submission(submission)
}
\arguments{
\item{submission}{character scalar. Submission Id.}
}
\value{
tibble. Information about deleted submission.
}
\description{
Delete this submission, including all the objects in it.
}
\examples{
\dontrun{
ega_delete_submission(submission=1)
}
}
