% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_enums.R
\name{ega_enums}
\alias{ega_enums}
\title{Show Predefined Values From EGA}
\usage{
ega_enums(enums = NULL)
}
\arguments{
\item{enums}{Character scalar. Show predefined values for the specified
enumeration. Default \code{NULL} shows all enumerations available.}
}
\value{
Tibble. Table with predefined values.
}
\description{
Show predefined values from EGA.
}
\examples{
\dontrun{
ega_enums()
}
}
