% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_ega.R
\name{ega_error_body}
\alias{ega_error_body}
\title{Print-out API Error Messages}
\usage{
ega_error_body(resp)
}
\arguments{
\item{resp}{\code{httr2_response} object}
}
\value{
Character Scalar. EGA API Error Message.
}
\description{
Print-out API Error Messages
}
\examples{
\dontrun{
resp <- last_response()
resp |> ega_error_body()
}
}
\keyword{internal}
