% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_experiments.R
\name{ega_experiments}
\alias{ega_experiments}
\title{List All Experiments in EGA}
\usage{
ega_experiments(submission = NULL, experiment = NULL)
}
\arguments{
\item{submission}{Get information for all experiments in this submission.}

\item{experiment}{Get information only for this experiment.}
}
\value{
tibble with information about experiments.
}
\description{
Get information about existing experiments in EGA. If provisional or permanent accession \code{experiment} id is specified, then only information about this experiment is shown.
}
\examples{
\dontrun{
ega_experiments()
}

}
