% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_files.R
\name{ega_files}
\alias{ega_files}
\title{Show Information About Existing Files In EGA}
\usage{
ega_files(submission = NULL, prefix = NULL, status = NULL)
}
\arguments{
\item{submission}{Character scalar. Get information for all files
in this submission ID.}

\item{prefix}{Character scalar. Show files matching this prefix.}

\item{status}{Character scalar. Show files matching this status.}
}
\value{
Tibble. Table with information about the files.
}
\description{
Show information about all the files for the current EGA user.
}
\examples{
\dontrun{
ega_files()
}
}
