% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_submissions.R
\name{ega_finalise_submission}
\alias{ega_finalise_submission}
\title{Finalise The Submission In EGA}
\usage{
ega_finalise_submission(
  submission,
  expected_release_date = Sys.Date() + 365,
  dataset_changelogs = NULL
)
}
\arguments{
\item{submission}{character scalar. Submission Id.}

\item{expected_release_date}{character scalar. Expected release date in the format YYYY-MM-DD.}

\item{dataset_changelogs}{list. List with two elements: \code{dataset} with dataset ids and \code{message} change log messages.}
}
\value{
tibble. Information about finalised submission.
}
\description{
Finalise this submission.
}
\examples{
\dontrun{
ega_finalise_submission(submission=1)
}
}
