% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_get.R
\name{ega_get}
\alias{ega_get}
\title{Main EGA Get Function}
\usage{
ega_get(resource_prefix, resource_id = NULL, resource_suffix = NULL)
}
\arguments{
\item{resource_prefix}{character scalar. First part of API url}

\item{resource_id}{character scalar. EGA object ID}

\item{resource_suffix}{character scalar.}
}
\value{
tibble with obtained information.
}
\description{
Main EGA Get Function
}
\examples{
\dontrun{
ega_get(resource_prefix="enums")
}

}
\keyword{internal}
