% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_ega.R
\name{ega_get_password}
\alias{ega_get_password}
\title{Get EGA User Password}
\usage{
ega_get_password()
}
\value{
character scalar with EGA password
}
\description{
\code{ega_get_password} get EGA password either from ENV variable "REGA_EGA_PASSWORD" or ask the user to provide it.
The encrypted password is stored in the ENV variable "REGA_EGA_PASSWORD". ENV variable "REGA_KEY" is used to en-/de-crypt the password.
}
\examples{
ega_get_password <- Rega:::ega_get_password()

}
\keyword{internal}
