% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_ega.R
\name{ega_get_username}
\alias{ega_get_username}
\title{Get EGA Username}
\usage{
ega_get_username()
}
\value{
character scalar with EGA username
}
\description{
\code{ega_get_username} get EGA username either from ENV variable "REGA_EGA_USERNAME" or ask the user to provide it.
}
\examples{
ega_get_username <- Rega:::ega_get_username()

}
\keyword{internal}
