% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_get.R
\name{ega_parse_body}
\alias{ega_parse_body}
\title{Parse The Information From httr2 Response Object.}
\usage{
ega_parse_body(resp)
}
\arguments{
\item{resp}{httr2 response object.}
}
\value{
tibble with obtained information.
}
\description{
Parse the information from the body of \code{httr2} response object and returns it as a \code{tibble}.
}
\examples{
\dontrun{
ega_parse_body(resp)
}
}
\keyword{internal}
