% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_policies.R
\name{ega_polices}
\alias{ega_polices}
\title{List All Polices In EGA}
\usage{
ega_polices(dac = NULL)
}
\arguments{
\item{dac}{Character scalar. Show existing polices only for this DAC.}
}
\value{
Tibble. Table with showing existing polices.
}
\description{
Get information about existing policies. If \code{DAC} is specified, then only
policies corresponding to the specific Data Access Committee (DAC) are shown.
}
\examples{
\dontrun{
ega_policies()
}
}
