% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_analyses.R
\name{ega_rollback_analysis}
\alias{ega_rollback_analysis}
\title{Rollback Run In EGA}
\usage{
ega_rollback_analysis(analysis)
}
\arguments{
\item{analysis}{character scalar. Run accession Id.}
}
\value{
tibble. Information about rollback of the analysis.
}
\description{
Undo all the changes done to this analysis.
}
\examples{
\dontrun{
ega_rollback_analysis(analysis=1)
}
}
