% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_datasets.R
\name{ega_rollback_dataset}
\alias{ega_rollback_dataset}
\title{Rollback Dataset In EGA}
\usage{
ega_rollback_dataset(dataset)
}
\arguments{
\item{dataset}{character scalar. Dataset accession Id.}
}
\value{
tibble. Information about rollback of the dataset.
}
\description{
Undo all the changes done to this dataset.
}
\examples{
\dontrun{
ega_rollback_dataset(dataset=1)
}
}
