% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_experiments.R
\name{ega_rollback_experiment}
\alias{ega_rollback_experiment}
\title{Rollback experiment In EGA}
\usage{
ega_rollback_experiment(experiment)
}
\arguments{
\item{experiment}{character scalar. experiment accession Id.}
}
\value{
tibble. Information about rollback of the experiment.
}
\description{
Undo all the changes done to this experiment.
}
\examples{
\dontrun{
ega_rollback_experiment(experiment=1)
}
}
