% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_runs.R
\name{ega_rollback_run}
\alias{ega_rollback_run}
\title{Rollback Run In EGA}
\usage{
ega_rollback_run(run)
}
\arguments{
\item{run}{character scalar. Run accession Id.}
}
\value{
tibble. Information about rollback of the run.
}
\description{
Undo all the changes done to this run.
}
\examples{
\dontrun{
ega_rollback_run(run=1)
}
}
