% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_samples.R
\name{ega_rollback_sample}
\alias{ega_rollback_sample}
\title{Rollback Sample In EGA}
\usage{
ega_rollback_sample(sample)
}
\arguments{
\item{sample}{character scalar. Sample accession Id.}
}
\value{
tibble. Information about rollback of the sample.
}
\description{
Undo all the changes done to this sample.
}
\examples{
\dontrun{
ega_rollback_sample(sample=1)
}
}
