% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_studies.R
\name{ega_rollback_study}
\alias{ega_rollback_study}
\title{Rollback study In EGA}
\usage{
ega_rollback_study(study)
}
\arguments{
\item{study}{character scalar. Study accession Id.}
}
\value{
tibble. Information about rollback of the study.
}
\description{
Undo all the changes done in all the objects in this study which have been modified, including the study itself, and delete any new object that has been created.
}
\examples{
\dontrun{
ega_rollback_study(study=1)
}
}
