% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_submissions.R
\name{ega_rollback_submission}
\alias{ega_rollback_submission}
\title{Rollback Submission In EGA}
\usage{
ega_rollback_submission(submission)
}
\arguments{
\item{submission}{character scalar. Submission Id.}
}
\value{
tibble. Information about rollback of the submission.
}
\description{
Undo all the changes done in all the objects in this submission which have been modified, including the submission itself, and delete any new object that has been created.
}
\examples{
\dontrun{
ega_rollback_submission(submission=1)
}
}
