% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_runs.R
\name{ega_runs}
\alias{ega_runs}
\title{List All Runs in EGA}
\usage{
ega_runs(submission = NULL, run = NULL)
}
\arguments{
\item{submission}{Get information for all runs in this submission.}

\item{run}{Get information only for this run.}
}
\value{
tibble with information about runs.
}
\description{
Get information about existing runs in EGA. If provisional or permanent accession \code{run} id is specified, then only information about this run is shown.
}
\examples{
\dontrun{
ega_runs()
}

}
