% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_samples.R
\name{ega_samples}
\alias{ega_samples}
\title{List All Samples in EGA}
\usage{
ega_samples(submission = NULL, sample = NULL)
}
\arguments{
\item{submission}{Get information for all samples in this submission.}

\item{sample}{Get information only for this sample.}
}
\value{
tibble with information about samples.
}
\description{
Get information about existing samples in EGA. If provisional or permanent accession \code{sample} id is specified, then only information about this sample is shown.
}
\examples{
\dontrun{
ega_samples()
}

}
