% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_studies.R
\name{ega_studies}
\alias{ega_studies}
\title{List All Studies in EGA}
\usage{
ega_studies(submission = NULL, study = NULL)
}
\arguments{
\item{submission}{Get information for all studies in this submission.}

\item{study}{Get information only for this study.}
}
\value{
tibble with information about studies.
}
\description{
Get information about existing studies in EGA. If provisional or permanent accession \code{study} id is specified, then only information about this study is shown.
}
\examples{
\dontrun{
ega_studies()
}

}
