% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_submissions.R
\name{ega_submissions}
\alias{ega_submissions}
\title{List All Submissions In EGA}
\usage{
ega_submissions(submission = NULL)
}
\arguments{
\item{submission}{Get information only for this submission.}
}
\value{
tibble with information about submission(s).
}
\description{
Get information about existing submissions in EGA for the current user. If provisional or permanent accession \code{submission} id is specified, then only information about this submission is shown.
}
\examples{
\dontrun{
ega_submissions()
}

}
