% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{ega_token}
\alias{ega_token}
\title{Retrieve EGA API Bearer Token}
\usage{
ega_token(
  username = .get_ega_username(),
  password = .get_ega_password(),
  token_url = .EGA_TOKEN_URL
)
}
\arguments{
\item{username}{Character. The username for EGA authentication. Defaults to
the value returned by \code{.get_ega_username()}.}

\item{password}{Character. The password for EGA authentication. Defaults to
the value returned by \code{.get_ega_password()}.}

\item{token_url}{Character. The URL for the EGA token endpoint. Defaults to
the standard EGA token URL if not provided. Defaults to .EGA_TOKEN_URL =
\code{"https://idp.ega-archive.org/realms/EGA/protocol/openid-connect/token"}.}
}
\value{
A list containing the token details if successful. Actual token value
can be retrieved by \code{token$access_token}
}
\description{
This function retrieves an API token from the European Genome-Phenome Archive
(EGA) using user credentials.
}
\examples{
try(
    ega_token(username = "my_username", password = "my_password")
)

try(
    ega_token(token_url = "https://www.example.com")
)

}
