% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_analyses.R
\name{ega_update_analysis}
\alias{ega_update_analysis}
\title{Update Analysis In EGA}
\usage{
ega_update_analysis(
  analysis,
  title,
  description,
  analysis_type,
  files,
  experiment_types = NULL,
  genome_id = NULL,
  chromosomes = NULL,
  platform = NULL,
  study_provisional_id = NULL,
  study_accession_id = NULL,
  experiment_provisional_ids = NULL,
  experiment_accession_ids = NULL,
  sample_provisional_ids = NULL,
  sample_accession_ids = NULL,
  extra_attributes = NULL
)
}
\arguments{
\item{analysis}{character scalar. analysis Id.}

\item{title}{Character scalar. The analysis title.}

\item{description}{Character scalar. The analysis description.}

\item{analysis_type}{Character scalar. The analysis analysis type. See \code{ega_enums("analysis_types")}.}

\item{files}{List. Provisional Ids of the files.}

\item{experiment_types}{Character scalar. Types of experiments.}

\item{genome_id}{Integer scalar. The analysis genome id. See \code{ega_enums("genomes")}.}

\item{chromosomes}{List. See \code{ega_enums("chromosomes")}.}

\item{platform}{Character scalar. The analysis platform.}

\item{study_provisional_id}{Integer scalar. The analysis's study's provisional ID.}

\item{study_accession_id}{Character scalar. The analysis's study's accession ID.}

\item{experiment_provisional_ids}{Integer scalar. The analysis's experiment's provisional ID.}

\item{experiment_accession_ids}{Character scalar. The analysis's experiment's accession ID.}

\item{sample_provisional_ids}{Integer vector. The analysis's sample's accession ID}

\item{sample_accession_ids}{Character scalar. The analysis's sample's accession ID.}

\item{extra_attributes}{List. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}
}
\value{
tibble. Information about updated analysis.
}
\description{
Update information of this analysis.
}
\examples{
\dontrun{
ega_update_analysis(analysis=1,
alias="My unique analysis alias 1",
biological_sex="male",
phenotype="nose",
subject_id="192873366738836788")
}
}
