% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_datasets.R
\name{ega_update_dataset}
\alias{ega_update_dataset}
\title{Update Dataset In EGA}
\usage{
ega_update_dataset(
  dataset,
  title,
  description,
  dataset_types,
  policy_accession_id,
  run_provisional_ids = NULL,
  run_accession_ids = NULL,
  analysis_provisional_ids = NULL,
  analysis_accession_ids = NULL,
  extra_attributes = NULL
)
}
\arguments{
\item{dataset}{character scalar. dataset Id.}

\item{title}{Character scalar. Dataset title.}

\item{description}{Character scalar. Dataset description.}

\item{dataset_types}{Character scalar. The dataset types. See \code{ega_enums("dataset_types")}.}

\item{policy_accession_id}{Character scalar. The policy accession ID.}

\item{run_provisional_ids}{Integer vector. Provisional Run IDs.}

\item{run_accession_ids}{Character vector. Accession Run IDs.}

\item{analysis_provisional_ids}{Integer vector. Provisional Analysis IDs.}

\item{analysis_accession_ids}{Character vector. Accession Analysis IDs.}

\item{extra_attributes}{List. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}
}
\value{
tibble. Information about updated dataset.
}
\description{
Update information of this dataset.
}
\examples{
\dontrun{
ega_update_dataset(dataset=1,
alias="My unique dataset alias 1",
biological_sex="male",
phenotype="nose",
subject_id="192873366738836788")
}
}
