% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_runs.R
\name{ega_update_run}
\alias{ega_update_run}
\title{Update Run In EGA}
\usage{
ega_update_run(
  run,
  run_file_type,
  files,
  experiment_provisional_id = NULL,
  experiment_accession_id = NULL,
  sample_provisional_id = NULL,
  sample_accession_id = NULL,
  extra_attributes = NULL
)
}
\arguments{
\item{run}{character scalar. run Id.}

\item{run_file_type}{Character scalar. File type.}

\item{files}{List. Provisional Ids of the files.}

\item{experiment_provisional_id}{Character scalar. Provisional Experiment Id.}

\item{experiment_accession_id}{Character scalar. Experiment Accession Id.}

\item{sample_provisional_id}{Character scalar. Provisional Sample Id.}

\item{sample_accession_id}{character scalar. Sample Accession Id.}

\item{extra_attributes}{List. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}
}
\value{
tibble. Information about updated run.
}
\description{
Update information of this run.
}
\examples{
\dontrun{
ega_update_run(run=1,
alias="My unique run alias 1",
biological_sex="male",
phenotype="nose",
subject_id="192873366738836788")
}
}
