% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_samples.R
\name{ega_update_sample}
\alias{ega_update_sample}
\title{Update sample In EGA}
\usage{
ega_update_sample(
  sample,
  alias,
  biological_sex,
  subject_id,
  phenotype,
  title = NULL,
  description = NULL,
  biosample_id = NULL,
  case_control = NULL,
  organism_part = NULL,
  cell_line = NULL,
  extra_attributes = NULL
)
}
\arguments{
\item{sample}{character scalar. Sample Id.}

\item{alias}{character scalar. Sample alias (must be unique within the study).}

\item{biological_sex}{character scalar. Biological sex of subject, for accepted values check \code{ega_enums("biological_sex")}.}

\item{subject_id}{character scalar. Subject Id.}

\item{phenotype}{character scalar. Phenotype group}

\item{title}{character scalar. Sample title.}

\item{description}{character scalar. Sample description}

\item{biosample_id}{character scalar. Biosample Id.}

\item{case_control}{character scalar. Case or Control, for accepted values check \code{ega_enums("case_controls")}.}

\item{organism_part}{character scalar. Sample tissue or cell type.}

\item{cell_line}{character scalar. Name of teh cell line}

\item{extra_attributes}{list. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}
}
\value{
tibble. Information about updated sample.
}
\description{
Update information of this sample.
}
\examples{
\dontrun{
ega_update_sample(sample=1,
alias="My unique sample alias 1",
biological_sex="male",
phenotype="nose",
subject_id="192873366738836788")
}
}
