% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_studies.R
\name{ega_update_study}
\alias{ega_update_study}
\title{Update study In EGA}
\usage{
ega_update_study(
  study,
  title,
  description,
  study_type,
  pubmed_ids = NULL,
  custom_tags = NULL,
  extra_attributes = NULL,
  repositories = NULL
)
}
\arguments{
\item{study}{character scalar. study Id.}

\item{title}{character scalar. study title.}

\item{description}{character scalar. study description.}

\item{study_type}{Study type, for options check output of \code{ega_enums(enums="study_types")}.}

\item{pubmed_ids}{list. List study PubMedIDs. Single element list without names.}

\item{custom_tags}{list. The study custom tags. Single element list without names.}

\item{extra_attributes}{list. The extra attributes, list with element names: \code{tag}, \code{value} and \code{unit.}}

\item{repositories}{list. The external references to other repositories, list with element names: \code{repository_id}, \code{url} and \code{label}.}
}
\value{
tibble. Information about updated study.
}
\description{
Update information of this study.
}
\examples{
\dontrun{
ega_update_study(study=1, title="study Title (update)",
description="study description (update)")
}
}
