% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_submissions.R
\name{ega_update_submission}
\alias{ega_update_submission}
\title{Update Submission In EGA}
\usage{
ega_update_submission(submission, title, description)
}
\arguments{
\item{submission}{character scalar. Submission Id.}

\item{title}{character scalar. Submission title.}

\item{description}{character scalar. Submission description.}
}
\value{
tibble. Information about updated submission.
}
\description{
Update information of this submission.
}
\examples{
\dontrun{
ega_update_submission(submission=1, title="Submission Title (update)",
description="Submission description (update)")
}
}
