% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ega_users.R
\name{ega_users}
\alias{ega_users}
\title{Show User ID From EGA}
\usage{
ega_users(username = ega_get_username())
}
\arguments{
\item{username}{Character scalar. EGA username.}
}
\value{
Tibble. Table with username and user ID.
}
\description{
Show current user ID for the specified username.
}
\examples{
\dontrun{
ega_userid()
}

}
