% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{file_formatter}
\alias{file_formatter}
\title{Format File Table with EGA File Paths}
\usage{
file_formatter(tab, params)
}
\arguments{
\item{tab}{Data frame. The input table containing file information. Columns
\code{file}, \code{ega_inbox_relative_path} need to be present in the data.}

\item{params}{List. Additional parameters for formatting. Takes a formatter
params value from parser parameter yaml file. Includes \code{crypt_ext} for
encryption file extensions and \code{prepend_slash} to control path prefix.}
}
\value{
A formatted data frame with cleaned column names, and updated
\code{ega_file} paths based on file and relative path information.
}
\description{
Format File Table with EGA File Paths
}
\examples{
params <- list(prefix = "", crypt_ext = "c4gh", prepend_slash = FALSE)

# Dummy data, first row will be moved to column names
tab <- data.frame(
    x1 = c("file", "value1", "value2"),
    x2 = c("ega_inbox_relative_path", NA, "proj1")
)

file_formatter(tab, params)

}
