% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_helpers.R
\name{files_in_inbox}
\alias{files_in_inbox}
\title{Check whether all expected files are in the inbox}
\usage{
files_in_inbox(file_list, client = NULL)
}
\arguments{
\item{file_list}{A character vector or list of file prefixes to check.}

\item{client}{List of functions. EGA API client created by \code{create_client}
function from EGA API schema. If \code{NULL}, default client will be created by
\code{create_client(extract_api())}. Defaults to \code{NULL}.}
}
\value{
A logical scalar, \code{TRUE} if all files are present, otherwise
\code{FALSE}.
}
\description{
Query the remote client for each requested file prefix and test whether a
file is found for every element of \code{file_list}.
}
\examples{
files_in_inbox(list("file_a", "file_b"), client)

}
