% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{finalise_submission}
\alias{finalise_submission}
\title{Finalise an EGA submission}
\usage{
finalise_submission(
  id,
  release_date,
  dataset_changelogs = data.frame(),
  client = NULL,
  logfile = NULL,
  ...
)
}
\arguments{
\item{id}{Character scalar. The submission accession or provisional ID.}

\item{release_date}{Character scalar. Expected release date in YYYY-MM-DD
format.}

\item{dataset_changelogs}{Data frame. Optional changelog metadata for
associated datasets. If specified, the requred columns are \code{dataset} and
\code{message}. Defaults to empty data.frame.}

\item{client}{List of functions. EGA API client created by \code{create_client}
function from EGA API schema. If \code{NULL}, default client will be
created by \code{create_client(extract_api())}. Defaults to \code{NULL}.}

\item{logfile}{Character. Path of log file to log the \code{httr2} responses from
individual operations or \code{NULL}. Defaults to \code{NULL}.}

\item{...}{List. Additional arguments to the function.}
}
\value{
The API response object from the finalisation request.
}
\description{
Submits the finalisation request for a submission identified by either an
accession or provisional ID. Validates the release date and sends optional
dataset changelogs.
}
\examples{
# Requires credentials
try(
    finalise_submission("123456", "2025-12-31")
)

}
