% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{first_row_to_colnames}
\alias{first_row_to_colnames}
\title{Use First Row as Column Names for a Data Frame}
\usage{
first_row_to_colnames(df, to_api = TRUE)
}
\arguments{
\item{df}{Data frame. The input data frame whose first row will become column
names.}

\item{to_api}{Logical. Whether to convert labels to API-style names using
\code{label_to_api_name()}. Defaults to \code{TRUE}.}
}
\value{
A data frame with updated column names and the first row removed.
}
\description{
Use First Row as Column Names for a Data Frame
}
\examples{
df <- data.frame(id = c("A B", "C D_"), value = c("* E F", "GH"))

first_row_to_colnames(df)
first_row_to_colnames(df, to_api = FALSE)

}
