% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_chr_group}
\alias{get_chr_group}
\title{Retrieve Chromosome Belonging to a Group}
\usage{
get_chr_group(group_id, chr_enum, sep = "--")
}
\arguments{
\item{group_id}{Character. The group ID to filter by.}

\item{chr_enum}{Character vector. Chromosome enumeration data, where each
element is a string containing fields separated by \code{sep}. Enum data is
created by \code{parse_enum} and \code{get_enum} functions. See vignette for more
details.}

\item{sep}{Character. Field separator in a string. Defaults to \code{"--"}}
}
\value{
An integer vector of chromosome IDs corresponding to the specified
group ID.
}
\description{
Retrieve Chromosome Belonging to a Group
}
\examples{
get_chr_group(
 "group1", c("group1--1--chr1--name1", "group2--2--chr2--name2")
)

}
