% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{get_ega_password}
\alias{get_ega_password}
\title{Get EGA User Password}
\usage{
get_ega_password()
}
\value{
character scalar with EGA password
}
\description{
\code{get_ega_password} get EGA password either from ENV variable "REGA_EGA_PASSWORD"
or ask the user to provide it. The encrypted password is stored in the ENV
variable "REGA_EGA_PASSWORD". ENV variable "REGA_KEY" is used to en-/de-crypt
the password.
}
\examples{
ega_get_password <- Rega:::get_ega_password()

}
\keyword{internal}
