% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{get_ega_username}
\alias{get_ega_username}
\title{Get EGA Username}
\usage{
get_ega_username()
}
\value{
character scalar with EGA username
}
\description{
\code{get_ega_username} get EGA username either from ENV variable
"REGA_EGA_USERNAME" or ask the user to provide it.
}
\examples{
ega_get_username <- Rega:::get_ega_username()

}
\keyword{internal}
