% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_entry_by_title}
\alias{get_entry_by_title}
\title{Retrieve EGA entries by title}
\usage{
get_entry_by_title(title, type = NULL, client = NULL, logfile = NULL, ...)
}
\arguments{
\item{title}{Character scalar. The title or substring to search for.}

\item{type}{Character vector. One or more metadata types ("submissions",
"studies", "samples", "experiments", "runs", "analyses" and "datasets"). If
NULL, searches all valid types.}

\item{client}{List of functions. EGA API client created by \code{create_client}
function from EGA API schema. If \code{NULL}, default client will be
created by \code{create_client(extract_api())}. Defaults to \code{NULL}.}

\item{logfile}{Character. Path of log file to log the \code{httr2} responses from
individual operations or \code{NULL}. Defaults to \code{NULL}.}

\item{...}{List. Additional arguments to the function.}
}
\value{
A named list of data frames containing entries matching the title.
}
\description{
Searches for entries across specified EGA metadata types that match a given
title string. Returns a list of data frames for each type.
}
\examples{
# Requires credentials
try(
    get_entry_by_title("My Study", type = "studies")
)

}
