% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_formatter}
\alias{get_formatter}
\title{Retrieve a Formatter Function by Type of Submission Metadata Table}
\usage{
get_formatter(x, params)
}
\arguments{
\item{x}{Character. The name of the submission metadata table/sheet.}

\item{params}{List. A list containing a \code{formatter} element from parser
parameter yaml file.}
}
\value{
The formatter function corresponding to the specified table.
}
\description{
Retrieve a Formatter Function by Type of Submission Metadata Table
}
