% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_formatter_params}
\alias{get_formatter_params}
\title{Retrieve Formatter Parameters by Name}
\usage{
get_formatter_params(x, params)
}
\arguments{
\item{x}{Character. The name of the formatter for which to retrieve
parameters.}

\item{params}{List. A list containing a \code{formatter} element from parser
parameter yaml file.}
}
\value{
A list of parameters for the specified formatter.
}
\description{
Retrieve Formatter Parameters by Name
}
