% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{get_operation_schema}
\alias{get_operation_schema}
\title{Retrieve the Schema for an API Operation}
\usage{
get_operation_schema(op)
}
\arguments{
\item{op}{List. The API operation definition containing a \code{requestBody}
element with content and schema details.}
}
\value{
The schema for the operation's JSON request body, or \code{NULL} if
no schema is defined.
}
\description{
Retrieve the Schema for an API Operation
}
\examples{
# Get operations from API
opdefs <- extract_operation_definitions(extract_api())

# Retrieve the schema for a specific operation
schema <- get_operation_schema(opdefs[["post__submissions"]])

}
