% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_schemas}
\alias{get_schemas}
\title{Retrieve Request schemas from an API Specification}
\usage{
get_schemas(api)
}
\arguments{
\item{api}{List. The API specification, typically containing a \code{components}
element with nested \code{schemas}.}
}
\value{
A list of schemas whose names contain "Request", filtered from the
\code{schemas} element of the API specification.
}
\description{
This function extracts and returns schemas related to requests from the API
specification.
}
\examples{
# Extract request schemas from an API specification
request_schemas <- get_schemas(extract_api())

}
