% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sentence_number}
\alias{get_sentence_number}
\title{Get number of sentences in the text}
\usage{
get_sentence_number(text)
}
\arguments{
\item{text}{}
}
\value{
Integer, number of sentences in the text
}
\description{
Get number of sentences in the text
}
