% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_word_number}
\alias{get_word_number}
\title{Count the number of words in text}
\usage{
get_word_number(text)
}
\arguments{
\item{text}{A character vector with text.}
}
\value{
An integer vector giving the number of words per element of
\code{text}.
}
\description{
Compute the number of words in each element of a character vector using
non-word separators.
}
\examples{
get_word_number(c("one two", "three four five"))

}
