% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{has_linked_sheets}
\alias{has_linked_sheets}
\title{Check for Linked Sheets in Metadata}
\usage{
has_linked_sheets(metadata, colname)
}
\arguments{
\item{metadata}{A list of data frame objects to check.}

\item{colname}{A string specifying the name of the column to look for.}
}
\value{
A logical vector indicating whether each element of \code{metadata}
contains the specified column with at least one non-NA value.
}
\description{
Determines whether a specified sheet is present and contains at least one
non-NA value in the provided metadata.
}
\examples{
metadata <- list(
    sheet1 = list(sheet_name = c(1, NA)),
    sheet2 = list(other_name = NA)
)
has_linked_sheets(metadata, "sheet_name")

}
