% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_helpers.R
\name{is_provisional}
\alias{is_provisional}
\title{Check whether IDs is a provisional ID.}
\usage{
is_provisional(x)
}
\arguments{
\item{x}{A numeric or character vector of candidate provisional IDs.}
}
\value{
A logical vector indicating which values are provisional IDs.
}
\description{
Determine if input values match the format of provisional IDs, either as
whole-number numerics or as character strings of at least two digits without
leading zeros.
}
\examples{
is_provisional(c(10, 11, 3.5, 9))

}
