% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{is_valid_http_method}
\alias{is_valid_http_method}
\title{Validate HTTP Method}
\usage{
is_valid_http_method(m)
}
\arguments{
\item{m}{A string representing the HTTP method to validate.}
}
\value{
A logical value: \code{TRUE} if \code{m} is a valid HTTP method,
otherwise \code{FALSE}.
}
\description{
Checks whether a given HTTP method is valid based on a predefined list of
accepted methods (matches on lowercase).
}
\examples{
is_valid_http_method("GET") # TRUE
is_valid_http_method("get") # TRUE
is_valid_http_method("DELETE") # TRUE
is_valid_http_method("foo") # FALSE
is_valid_http_method(NULL) # FALSE
}
