% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{label_to_api_name}
\alias{label_to_api_name}
\title{Convert Prettified Labels to API Names}
\usage{
label_to_api_name(x, req_str = "* ")
}
\arguments{
\item{x}{Character vector. Prettified labels to convert.}

\item{req_str}{Character. Optional prefix to remove from labels. Defaults to
\code{"* "}.}
}
\value{
A character vector with labels converted to API-style names.
}
\description{
Convert Prettified Labels to API Names
}
\examples{
\dontrun{
# Convert labels to API names
api_names <- label_to_api_name(c("* First Name", "Last Name"))
}

}
