% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{link_sheet}
\alias{link_sheet}
\title{Link Data Between Metadata Sheets}
\usage{
link_sheet(metadata, sheet_name)
}
\arguments{
\item{metadata}{List. A list of data frames representing metadata sheets.}

\item{sheet_name}{Character. The name of the sheet to link with other sheets.}
}
\value{
A list of updated metadata with the specified values replaced based on
referenced values present in \code{sheet_name}.
}
\description{
Data frames representing metadata sheets that contain column names
corresponding to \code{sheet_name} containing an ID reference (to a first column
in \code{sheet_name}) will be replaced with the rest of the values nested as a list.
}
\examples{
\dontrun{
# Link data from a specific sheet to other sheets in metadata
metadata <- list(
  sheet1 = data.frame(id = c(1, 2), linked_sheet = c("A", "B")),
  linked_sheet = data.frame(id = c("A", "B"), value = c(10, 20))
)
updated_metadata <- link_sheet(metadata, "linked_sheet")
}

}
