% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{lut_add}
\alias{lut_add}
\title{Add a Column to a Data Frame Based on Lookup Table}
\usage{
lut_add(df, to, from, lut)
}
\arguments{
\item{df}{A data frame to which the new column will be added.}

\item{to}{A string specifying the name of the new column.}

\item{from}{A string specifying the name of the column to map values from.}

\item{lut}{A named list or vector serving as the lookup table.}
}
\value{
The input data frame with the added column.
}
\description{
This function adds a new column to a data frame by mapping values from an
existing column through a lookup table.
}
\examples{
df <- data.frame(id = c("A", "B", "C"), stringsAsFactors = FALSE)
lut <- list(A = 1, B = 2, C = 3)
lut_add(df, "value", "id", lut)

}
