% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{merge_linked_sheet}
\alias{merge_linked_sheet}
\title{Merge Linked Sheets with Source Data}
\usage{
merge_linked_sheet(target, source, dat, sheet)
}
\arguments{
\item{target}{A vector containing the target values.}

\item{source}{A string specifying the source column to merge on.}

\item{dat}{A data frame representing the data to be linked.}

\item{sheet}{A string specifying the name of the sheet, used for API-specific
processing.}
}
\value{
A data frame containing the merged data, or an empty list if the
target is entirely \code{NA}.
}
\description{
Merges a target column with a source column in a linked sheet's data,
processing it into a format suitable for JSON parsing. Includes API-specific
adjustments for certain data.
}
\examples{
target <- c(1, 2, 3)
source <- "id"
dat <- data.frame(id = c(1, 2, 3), value = c("A", "B", "C"))
merge_linked_sheet(target, source, dat, "collaborators")

}
