% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{multi_lut_add}
\alias{multi_lut_add}
\title{Add Multiple Lookup-Based Columns to a Data Frame}
\usage{
multi_lut_add(df, ...)
}
\arguments{
\item{df}{A data frame to which new columns will be added.}

\item{...}{A series of lists, each containing three elements: the name of
the new column (\code{to}), the name of the source column (\code{from}),
and the lookup table (\code{lut}). See \code{\link{lut_add}} for details.}
}
\value{
The input data frame with the added columns.
}
\description{
Adds multiple columns to a data frame by applying multiple lookup tables,
each defined by a set of arguments specifying the new column, the source
column, and the lookup table.
}
\examples{
df <- data.frame(id = c("A", "B", "C"), stringsAsFactors = FALSE)
lut1 <- list(A = 1, B = 2, C = 3)
lut2 <- list(A = "x", B = "y", C = "z")
multi_lut_add(df, list("value1", "id", lut1), list("value2", "id", lut2))

}
