% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_to_empty_list}
\alias{na_to_empty_list}
\title{Convert NA Values to Empty Lists}
\usage{
na_to_empty_list(l)
}
\arguments{
\item{l}{A list containing elements that may include \code{NA} values.}
}
\value{
A list where any \code{NA} values have been replaced with empty
lists.
}
\description{
Replaces \code{NA} values in a list with empty lists, preserving the original
structure of the list.
}
\examples{
input_list <- list(1, NA, "text", NA)
na_to_empty_list(input_list)

}
