% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_enum}
\alias{parse_enum}
\title{Parse Enum into a Formatted String}
\usage{
parse_enum(enum, sep = "--")
}
\arguments{
\item{enum}{Data frame or character vector. The enum to parse. If a data
frame, its rows are concatenated into strings. If a character vector, its
elements are joined with newlines.}

\item{sep}{Character. If enum has multiple fields, they will be pasted into a
single string using this separator. Defaults to \verb{--}}
}
\value{
A single string representing the parsed enum. Rows are joined by
newlines and multiple enum fields are joined by \code{sep}.
}
\description{
This function parses an enum, represented as a data frame or character
vector, into a formatted string for display or further use.
}
\examples{
# Parse an enum as a data frame
df_enum <- data.frame(key = c("A", "B"), value = c("1", "2"))
parse_enum(df_enum)

# Parse an enum as a character vector
vec_enum <- c("A", "B", "C")
parse_enum(vec_enum)

}
