% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{process_chromosomes}
\alias{process_chromosomes}
\title{Process a Vector or List of Chromosome Data}
\usage{
process_chromosomes(chr_data, select_input_data)
}
\arguments{
\item{chr_data}{A list containing chromosome-related information. Expected to
have items \code{chromosomes} (scalar, vector or list) and/or
\code{chromosome_groups} (scalar).}

\item{select_input_data}{A list containing look-up data for \code{chromosomes}.}
}
\value{
A data frame with chromosome \code{id} and \code{label} if chromosomes are
present. If only chromosome groups exist, returns the result of lookup
against the \code{select_input_data} with \code{get_chr_group()}. If neither are
present, returns an empty list.
}
\description{
This function processes chromosome data by extracting unique chromosome IDs
and labels or retrieving chromosome group information from a lookup where
applicable.
}
\examples{
select_input_data <- list(
    chromosomes = c("group1--1--chr1--name1", "group1--2--chr2--name2")
)

chr_data_1 <- list(
    chromosomes = list("group1--1--chr1--name1", "group2--3--chr3--name3"),
    chromosome_groups = NA_character_
)
process_chromosomes(chr_data_1, select_input_data)

chr_data_2 <- list(
    chromosomes = NA,
    chromosome_groups = "group1"
)

process_chromosomes(chr_data_2, select_input_data)

}
