% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{process_delimited_column}
\alias{process_delimited_column}
\title{Process Delimited Columns in Metadata}
\usage{
process_delimited_column(
  metadata,
  column_name,
  separator,
  converters = DELIM_CONVERTERS
)
}
\arguments{
\item{metadata}{List. A list of data frames representing metadata sheets.}

\item{column_name}{Character. The name of the column to process.}

\item{separator}{Character. The delimiter used to split column values.}

\item{converters}{Named list of functions. Specifies how to coerce columns
with delimited strings into specific types or values. Defaults to
\code{DELIM_CONVERTERS}}
}
\value{
A list of updated metadata with the specified column split into lists
based on the delimiter and trimmed.
}
\description{
The specified column name is searched for across all the data frames. If the
column is \code{pubmed_ids}, values are converted to integers.
}
\examples{
metadata <- list(
    sheet1 = data.frame(pubmed_ids = c("123; 456", "130; 789; 102", NA))
)

process_delimited_column(metadata, "pubmed_ids", ";")

}
