% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rega-package.R
\docType{package}
\name{Rega-package}
\alias{Rega}
\alias{Rega-package}
\title{Rega: R Interface to European Genome-Phenome Archive}
\description{
Rega provides interface to European Genome-Phenome Archive (EGA).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ivanek/Rega}
  \item Report bugs at \url{https://github.com/ivanek/Rega/issues}
}

}
\author{
\strong{Maintainer}: Robert Ivanek \email{robert.ivanek@unibas.ch} (\href{https://orcid.org/0000-0002-8403-056X}{ORCID})

}
\keyword{internal}
