% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_ega.R
\name{req_ega}
\alias{req_ega}
\title{Send Request to EGA API}
\usage{
req_ega(resource, ..., method = "GET")
}
\arguments{
\item{resource}{Character scalar. Specific resource to be queried.}

\item{...}{named parameters which can be send with POST method as a body of httr2 request (in a JSON format)}

\item{method}{Character scalar. HTTP method like "GET", "POST", "DELETE"}
}
\value{
An HTTP response from EGA API: an S3 list with class \code{httr2_request}.
}
\description{
\code{req_ega} sends an request to EGA API.
}
\examples{
\dontrun{
Rega:::req_ega("enums")
}

}
\keyword{internal}
