% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_ega.R
\name{req_ega_auth}
\alias{req_ega_auth}
\title{Set The OAUTH With EGA Username And Password}
\usage{
req_ega_auth(
  req,
  ega_username = ega_get_username(),
  ega_password = ega_get_password()
)
}
\arguments{
\item{req}{A httr2 request.}

\item{ega_username}{Character scalar. EGA User name. By default the content of ENV variable "REGA_EGA_USERNAME".}

\item{ega_password}{Character scalar.EGA user Password. By default the content of ENV variable "REGA_EGA_PASSWORD". Please avoid entering the password directly when calling this function as it will be captured by \code{.Rhistory}.}
}
\value{
returns a modified HTTP request that will use OAuth;
}
\description{
\code{req_ega_auth} implements the EGA OAuth resource owner password flow, as defined by Section 4.3 of RFC 6749.
It allows the user to supply their password once, exchanging it for an access token that can be cached locally.
}
\section{Functions}{
\itemize{
\item \code{req_ega_auth()}: Set OAUTH for EGA API.

}}
\examples{
\dontrun{
Rega:::req_ega_auth(req)
}

}
\keyword{internal}
