% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{row_table_formatter}
\alias{row_table_formatter}
\title{Format a Row Table}
\usage{
row_table_formatter(tab, params)
}
\arguments{
\item{tab}{Data frame. The input table from a sumbission metadata file}

\item{params}{List. Additional parameters for formatting. Takes a formatter
params value from parser parameter yaml file.}
}
\value{
A cleaned and formatted tibble with correctly organized rows and
columns, whitespace trimmed, and folding applied to specified columns.
}
\description{
Format a Row Table
}
\examples{
# Formatter parameters
params = list(fold = "extra_attributes")

# Sample data frame
df = data.frame(
  ...1 = c("* Study", "* Title", "Extra Attributes", "Extra Attributes"),
  ...2 = c("Study1", "Title1", "A", "B"),
  ...3 = c("* Study", "* Title", "Extra Attributes", NA),
  ...4 = c("Study2", "Title2", "C", NA)
)

row_table_formatter(df, params)

}
