% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_helpers.R
\name{save_log}
\alias{save_log}
\title{Save API Responses to a Log File}
\usage{
save_log(responses, logfile)
}
\arguments{
\item{responses}{A list of responses to be saved.}

\item{logfile}{A string specifying the path to the log file. If \code{NULL},
no file is written.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
This function saves a list of API responses to a specified log file in YAML
format.
}
\examples{
\dontrun{
responses <- list(status = "success", data = list(a = 1, b = "text"))
save_log(responses, "log.yaml")
}

}
