% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_ega.R
\name{set_rega_key}
\alias{set_rega_key}
\title{Generate Rega Key}
\usage{
set_rega_key()
}
\description{
\code{set_rega_key} generates an encryption key and set it as an ENV variable
"REGA_KEY". Alternative might be to set permanent key on the user-level in
the \code{.Renviron} (which can be easily open with \verb{usethis::edit_r_environ())}
}
\examples{
Rega:::set_rega_key()

}
\keyword{internal}
