% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_helpers.R
\name{submit_table}
\alias{submit_table}
\title{Submit a Data Frame to an API Endpoint Row by Row}
\usage{
submit_table(tab, id, endpoint_func)
}
\arguments{
\item{tab}{A data frame containing the data to be submitted.}

\item{id}{An EGA accession/provisional ID passed to the \code{endpoint_func}.}

\item{endpoint_func}{A function that handles the API request. It should
accept \code{id} and a JSON \code{body} as arguments.}
}
\value{
Data frame. A combined response object from the API.
}
\description{
This function iterates over rows of a data frame, submitting each row to a
specified API endpoint function, and combines the responses into a single
data structure.
}
\examples{
tab <- data.frame(a = 1:2, b = c("x", "y"), stringsAsFactors = FALSE)
mock_endpoint <- function(id, body) list(id = id, body = body)
submit_table(tab, "example_id", mock_endpoint)

}
