% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_helpers.R
\name{unbox_list}
\alias{unbox_list}
\title{Convert a List to an Unboxed JSON-Compatible Data Frame}
\usage{
unbox_list(l)
}
\arguments{
\item{l}{A list where all elements must have a length of 1.}
}
\value{
A data frame with unboxed elements, suitable for JSON conversion.
}
\description{
Converts a list into a single-row data frame with unboxed elements if all
elements have a length of 1. Otherwise, an error is raised.
}
\examples{
input_list <- list(a = 1, b = "text", c = TRUE)
unbox_list(input_list)

}
