% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_helpers.R
\name{unbox_row}
\alias{unbox_row}
\title{Convert a Data Frame Row to an Unboxed JSON Object}
\usage{
unbox_row(row)
}
\arguments{
\item{row}{A single row of a data frame.}
}
\value{
A JSON object with unboxed values for the input row.
}
\description{
This function converts a single row of a data frame into an unboxed JSON
object, effectively removing the array structure.
}
\examples{
row <- data.frame(a = 1, b = "text", stringsAsFactors = FALSE)[1, ]
unbox_row(row)

}
