% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_schema}
\alias{validate_schema}
\title{Validate a Payload Against a JSON Schema}
\usage{
validate_schema(payload, schema)
}
\arguments{
\item{payload}{The payload to validate against the schema. JSON string or
single row of data frame converted to JSON representation with \code{unbox_row}
function or a list with all items of length 1 converted to JSON
representation with \code{unbox_list} function.}

\item{schema}{List. The JSON schema defining the validation rules.}
}
\value{
Logical value indicating whether the payload is valid. If invalid,
the result includes an \code{errors} attribute detailing the validation errors.
}
\description{
Function handles \code{oneOf} directives in a way that it in a case of validation
fail, it displays the overall result of the validation as first and then it
tests separately against all \code{oneOf} sub schemas.
}
\examples{
schema <- list(
    type = "object",
    properties = list(
        id = list(type = "integer"),
        title = list(type = "string")
    ),
    required = c("id")
)

payload_true <- data.frame(id = c(12345), title = c("abcd"))
payload_false <- data.frame(id = c("12345"), title = c(0.355))

validate_schema(jsonlite::unbox(payload_true), schema)
validate_schema(jsonlite::unbox(payload_false), schema)

}
