% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{api_function_factory}
\alias{api_function_factory}
\title{Generate an API Function from Operation and Specification}
\usage{
api_function_factory(
  op,
  api,
  verbosity = 0,
  bearer_token = NULL,
  token_url = NULL
)
}
\arguments{
\item{op}{List. The API operation definition, including method, path,
parameters, and request body schema.}

\item{api}{List. The API specification, including host and global security
definitions.}

\item{verbosity}{Integer, optional, values 0-3. Indicates with which
verbosity level should the requests \code{httr2::req_perform} be performed.
Default: 0.}

\item{bearer_token}{Character, optional. The API bearer token for
authentication, will be included in the headers of the request. Defaults to
\code{NULL}}

\item{token_url}{Character, optional. Token endpoint URL from which to obtain
the access token. If \code{bearer_token} is specified, it will take
precedence. If \code{NULL}, URL
\code{"https://idp.ega-archive.org/realms/EGA/protocol/openid-connect/token"}
will be used. Defaults to \code{NULL}.}
}
\value{
A dynamically generated function that performs the specified API
operation. The function accepts arguments corresponding to operation
parameters and executes the request using \code{httr2}.
}
\description{
This function dynamically creates an API function based on a given operation
definition and API specification. The generated function handles URL
construction, parameter validation, request execution, and response parsing.
}
\examples{
api <- extract_api()
opdefs <- extract_operation_definitions(api)

# Generate an API function for a specific operation
f <- api_function_factory(
    opdefs[["get__files"]], api,
    bearer_token = "my_key"
)

# Call the generated function with parameters (requires credentials)
try(
    result <- f(status = "value1", prefix = "value2")
)

}
