% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{column_table_formatter}
\alias{column_table_formatter}
\title{Format a Column Table}
\usage{
column_table_formatter(tab, params)
}
\arguments{
\item{tab}{Data frame. The input table sumbission metadata file where the
first row contains column names.}

\item{params}{List. Additional parameters for formatting. Takes a formatter
params value from parser parameter yaml file. Currently unused.}
}
\value{
A cleaned data frame with column names set from the first row, empty
rows removed, and whitespace trimmed from all values.
}
\description{
Format a Column Table
}
\examples{
\dontrun{
# Format a column table
tab <- data.frame(Alias = c(" Name1 ", " Name2 ", NA), Value = c(1, 2, NA))
formatted_tab <- column_table_formatter(tab, params = list())
}

}
