% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{create_client}
\alias{create_client}
\title{Generate API Client Functions}
\usage{
create_client(api, ...)
}
\arguments{
\item{api}{List. The API specification, including operation definitions,
host, and global settings.}

\item{...}{List. List of additional arguments passed to
\code{api_function_factory}}
}
\value{
A named list of functions, where each function corresponds to an API
operation. The function names match the operation IDs from the
specification.
}
\description{
This function creates a named list of functions for interacting with an API,
based on its specification and operation definitions.
}
\examples{
client <- create_client(extract_api(), api_key = "my_key", verbosity = 1)

# Call an operation using the client (requires credentials)
try(
  result <- client$get__files(status = "value1", prefix = "value2")
)

}
