% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{default_parser}
\alias{default_parser}
\title{Parser for a Default EGA Excel Template}
\usage{
default_parser(metadata_file, param_file = NULL)
}
\arguments{
\item{metadata_file}{Character. Path to a default template xlsx file
containing the submission metadata information.}

\item{param_file}{Character. Path to a yaml file with parameters for parser.}
}
\value{
List of data frames or lists. Submission information parsed from the
xlsx file.
}
\description{
This function parses the \code{extdata/ega_full_template_v3.xlsx} using the
bundled parser parameter file in \code{extdata/default_parser_params.yaml} to
extract information for EGA submission into format that can be easily passed
into EGA API endpoints.
}
\examples{
default_parser(
    system.file("extdata/submission_example.xlsx", package = "Rega")
)

}
