% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{default_validator}
\alias{default_validator}
\title{Validator for Default Parser}
\usage{
default_validator(meta, aliases = NULL)
}
\arguments{
\item{meta}{List of data frames. Correspond to tables of EGA submission.}

\item{aliases}{List of lists. Aliases that should present in the
EGA tables.  If \code{NULL}, the function will attempt to locate it in the
\code{meta} parameter. Defaults to \code{NULL}}
}
\value{
Data frame. Validator object that includes all performed validations
and their statistics (number of passes, fails and NAs, or whether errors or
warnings were encountered during validation)
}
\description{
Used to validate internal consistency of submission metadata parsed using
the default parser. Performs several checks on EGA dataset for submission,
ensuring that aliases for studies, experiments, samples, runs, analyses and
datasets are  are properly linked, as they will be replaced with provisional
or accession IDs during submission process. Displays a success message if
all validation passed or a summary message if validation failed. In addition
it returns a data frame with validation details.
}
\examples{
minimal_metadata <- list(
    aliases = list(
        studies = "Study1", experiments = "Experiment1",
        datasets = "Dataset1", samples = "Sample1", runs = "Run1",
        analyses = "Analysis1"
    ),
    files = tibble::tibble(
        file = "raw.fastq.gz", ega_file = list("raw.fastq.gz.c4gh")
    ),
    submission = tibble::tibble(title = "Submission"),
    studies = tibble::tibble(
        study = "Study1", title = "Study Title",
        description = "Study Description",
        study_type = "Whole Genome Sequencing"
    ),
    samples = tibble::tibble(
        alias = "Sample1", phenotype = "wild-type",
        biological_sex = "female", subject_id = "ID1"
    ),
    experiments = tibble::tibble(
        study = "Study1", experiment = "Experiment1",
        design_description = "Experiment Design",
        library_selection = "RANDOM", instrument_model_id = 1L,
        library_layout = "SINGLE", library_strategy = "WGS",
        library_source = "GENOMIC"
    ),
    runs = tibble::tibble(
        run = "Run1", experiment = "Experiment1", run_file_type = "srf",
        alias = "Sample1", files = list("raw.fastq.gz.c4gh")
    ),
    datasets = tibble::tibble(
        dataset = "Dataset1", title = "Dataset Title",
        description = "Dataset Description",
        policy_accession_id = "EGAP00000000001",
        dataset_types = list("Whole genome sequencing"),
        runs = list("Run1")
    )
)

default_validator(minimal_metadata)

}
