% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{delete_submission_contents}
\alias{delete_submission_contents}
\title{Delete Submission Contents and Log Responses}
\usage{
delete_submission_contents(id, client, logfile = NULL, ...)
}
\arguments{
\item{id}{A string representing the submission identifier. Can be either an
accession or provisional ID.}

\item{client}{An API client object with \code{delete} methods.}

\item{logfile}{A string specifying the path to a log file. If \code{NULL},
no log is written.}

\item{...}{Additional arguments for future extensions (currently unused).}
}
\value{
A list of responses for the deletion of associated datasets,
analyses, runs, experiments, samples, and studies.
}
\description{
Deletes all data associated with a submission ID using the client and logs
the responses if a logfile is specified.
}
\examples{
mock_client <- list(
    "delete__submissions__provisional_id__datasets" =
        function(id) list(status = "deleted")
)
delete_submission_contents(5678901, mock_client)

}
