% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\docType{data}
\name{DELIM_CONVERTERS}
\alias{DELIM_CONVERTERS}
\title{Default Column Converters for EGA Metadata}
\format{
A named list:
\describe{
\item{pubmed_ids}{Coerces values to integer.}
}
}
\usage{
DELIM_CONVERTERS
}
\description{
A named list of functions used to coerce delimited strings into
specific data types during metadata processing.
}
\examples{
df <- list(pub_sheet = data.frame(
    pubmed_ids = c("  123456,  789", NA, "1001  ")
))

process_delimited_column(
    df, "pubmed_ids", separator = ",", converters = DELIM_CONVERTERS
)

}
\keyword{datasets}
