% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{.add_headers}
\alias{.add_headers}
\title{Generate Header Expressions for an API Request}
\usage{
.add_headers(header_params, operation, api, token = NULL)
}
\arguments{
\item{header_params}{A character vector of header parameter names to include
in the request. Header parameters are created \code{.get_operation_params}
function.}

\item{operation}{List. The operation definition, which may include security
details.}

\item{api}{List. The API definition, which may include global security
details and other metadata.}
}
\value{
An expression to add headers to an API request using \code{req_headers()}.
}
\description{
This function creates expressions to add headers to an API request, including
content type, authorization, and any additional headers specified in the
parameters.
}
\examples{
api <- extract_api()
opdefs <- extract_operation_definitions(api)
params <- Rega:::.get_operation_params(opdefs[["get__files"]])
# No header parameters in operation, `Content-Type` added by default
Rega:::.add_headers(params$header, opdefs[["get__files"]], api)

}
\keyword{internal}
