% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{.add_json_validation}
\alias{.add_json_validation}
\title{Add JSON Schema Validation to API Operation}
\usage{
.add_json_validation(op)
}
\arguments{
\item{op}{A list representing the API operation, which may contain a
request body and schema.}
}
\value{
A list of expressions for JSON schema validation, or an empty list
if no schema is found.
}
\description{
Generates validation expressions for an API operation based on its JSON
schema. If a schema is present, the function returns expressions to validate
the request body and raise an error if validation fails.
}
\examples{
\dontrun{
op <- list(requestBody = TRUE, schema = list(type = "object"))
Rega:::.add_json_validation(op)
}

}
\keyword{internal}
