% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{.add_queries}
\alias{.add_queries}
\title{Generate Query Expressions for an API Request}
\usage{
.add_queries(query_params)
}
\arguments{
\item{query_params}{A character vector of query parameter names to include in
the request. Query parameters are created \code{.get_operation_params} function.}
}
\value{
An expression to add query parameters to an API request using
\code{req_url_query()}.
}
\description{
This function creates an expression to add query parameters to an API
request.
}
\examples{
Rega:::.add_queries("queries")

opdefs <- extract_operation_definitions(extract_api())
params <- Rega:::.get_operation_params(opdefs[["get__files"]])
Rega:::.add_queries(params$query)

}
\keyword{internal}
