% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{.add_queries}
\alias{.add_queries}
\title{Generate Query Parameter Expressions for an API Request}
\usage{
.add_queries(params)
}
\arguments{
\item{params}{List. A list containing \code{query}, a character vector of query
parameter names to include in the request.}
}
\value{
An expression to add query parameters to an API request using
\code{req_url_query()}.
}
\description{
This function creates an expression to add query parameters to an API
request.
}
\examples{
\dontrun{
# Generate query parameter expressions for a request
query_expr <- Rega:::.add_queries(params)
}

}
\keyword{internal}
