% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.analyses_extra_validator}
\alias{.analyses_extra_validator}
\title{Validate Analyses Metadata}
\usage{
.analyses_extra_validator(meta, aliases)
}
\arguments{
\item{meta}{A list containing parsed EGA metadata, where elements correspond
to individual sheets.}

\item{aliases}{A list of aliases, reference values used for validation.}
}
\value{
A validation object containing the results of the check or NULL if
analyses sheet was not present in metadata
}
\description{
Validates the metadata by checking the following:
\itemize{
\item analyses titles are unique
\item analyses descriptions are unique
\item files for analyses are specified
\item experiments for analyses are present in aliases
\item samples for analyses are present in aliases
}
}
\examples{
meta <- list(analyses = data.frame(
    title = c("TitleA", "TitleB", NA),
    description = c("DescriptionA", "DescriptionB", "DescriptionA"),
    samples = I(
        list(c("Sample1" , NA), c("Sample2", "Sample3"), c("Sample3"))
    ),
    experiments = I(list(NA, c("ExpA", "ExpB"), c("ExpD"))),
    files = I(list(c("File1", "File2"), c("File3"), NA))
))

aliases = list(
    samples = c("Sample1", "Sample3"),
    experiments = c("ExpA", "ExpB", "ExpC")
)

validate::summary(Rega:::.analyses_extra_validator(meta, aliases))

}
\keyword{internal}
