% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.datasets_extra_validator}
\alias{.datasets_extra_validator}
\title{Validate Datasets Metadata}
\usage{
.datasets_extra_validator(meta, aliases)
}
\arguments{
\item{meta}{A list containing parsed EGA metadata, where elements correspond
to individual sheets.}

\item{aliases}{A list of aliases, reference values used for validation (not
used in this function but included for consistency with other validators).}
}
\value{
A validation object containing the results of the check.
}
\description{
Validates the metadata by checking the following:
\itemize{
\item datasets titles are unique
\item datasets descriptions are unique
\item runs for datasets are present in aliases (as nested lists)
\item all runs in aliases are present in datasets (as nested lists)
}
}
\examples{
meta <- list(datasets = data.frame(
    title = c("TitleA", "TitleB", NA),
    description = c("DescriptionA", "DescriptionB", "DescriptionA"),
    runs = I(list(c("Run1" , NA), c("Run2", "Run3"), c("Run3")))
))

aliases = list(
    runs = c("Run1", "Run3", "Run4")
)

validate::summary(Rega:::.datasets_extra_validator(meta, aliases))

}
\keyword{internal}
