% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{.get_ega_password}
\alias{.get_ega_password}
\title{Retrieve EGA Password}
\usage{
.get_ega_password(keyring_name = "REGA_EGA", envvar = "REGA_EGA_PASSWORD", ...)
}
\arguments{
\item{keyring_name}{Character. Name of the service in the keyring. Defaults
to \code{"REGA_EGA"}.}

\item{envvar}{Character. Name of the environment variable. Defaults to
\code{"REGA_EGA_PASSWORD"}.}

\item{...}{Additional arguments passed to \code{.get_rega_key}.}
}
\value{
A character string containing the password.
}
\description{
Fetches the EGA password from the system keyring, an environment variable, or
an interactive prompt, following that priority order. If using the
environment variable, the value is decrypted using a key retrieved via
\code{.get_rega_key}.
}
\examples{
try(
    ega_password <- Rega:::.get_ega_password()
)

}
\keyword{internal}
