% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{.get_ega_password}
\alias{.get_ega_password}
\title{Get EGA User Password}
\usage{
.get_ega_password(envvar = "REGA_EGA_PASSWORD", ...)
}
\arguments{
\item{envvar}{A string specifying the name of the environment variable to
retrieve the password from. Defaults to \code{"REGA_EGA_PASSWORD"}.}

\item{...}{Additional arguments passed to \code{.get_rega_key}.}
}
\value{
A string containing the decrypted EGA password.
}
\description{
Retrieves the EGA password from the specified environment variable. If not
found, prompts the user to enter it. If an encryption key (usually stored in
REGA_KEY environmental variable) is available, the password is decrypted.
}
\examples{
ega_password <- Rega:::.get_ega_password()

}
\keyword{internal}
