% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{.get_ega_username}
\alias{.get_ega_username}
\title{Get EGA User Name}
\usage{
.get_ega_username(envvar = "REGA_EGA_USERNAME")
}
\arguments{
\item{envvar}{A string specifying the name of the environment variable to
retrieve the username from. Defaults to \code{"REGA_EGA_USERNAME"}.}
}
\value{
A string containing the EGA username.
}
\description{
Retrieves the EGA username from the specified environment variable. If the
variable is not set, prompts the user to enter their username interactively.
}
\examples{
ega_username <- Rega:::.get_ega_username()
ega_username <- Rega:::.get_ega_username("EGA_USERNAME")

}
\keyword{internal}
