% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{.get_rega_key}
\alias{.get_rega_key}
\title{Retrieve Rega secret from Environment Variable}
\usage{
.get_rega_key(envvar = "REGA_KEY")
}
\arguments{
\item{envvar}{A string specifying the name of the environment variable to
retrieve the API key from. Defaults to \code{"REGA_KEY"}.}
}
\value{
A string containing the REGA API key, or an empty string if the
variable is not set.
}
\description{
Retrieves the Rega secret from the specified environment variable. If the key
is not found, a warning message is issued. REGA secret should be generated
via \code{httr2::secret_make_key()} and stored as environmental variable either by
using \code{export} command in bash or at the the user-level in the \code{.Renviron}
file
}
\examples{
rega_key <- Rega:::.get_rega_key()

}
\keyword{internal}
