% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_client}
\alias{.is_client}
\title{Validate that an object is an API client.}
\usage{
.is_client(client)
}
\arguments{
\item{client}{An object to be tested, expected to be a list of functions.}
}
\value{
Returns \code{TRUE} if the validation succeeds, otherwise stops
execution with an error.
}
\description{
Checks that \code{client} is a list where every element is a function.
Additionally, it verifies that the name of each function starts with a valid
HTTP method (e.g., \code{get_}, \code{post_}, etc.). It stops execution with
an error if any of the checks fail.
}
\examples{
# Returns TRUE silently for a valid client structure
valid_client <- list(
    get_resource = function(...) "GET",
    post_data = function(...) "POST"
)
Rega:::.is_client(valid_client)

}
\keyword{internal}
