% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.studies_extra_validator}
\alias{.studies_extra_validator}
\title{Validate Studies Metadata}
\usage{
.studies_extra_validator(meta, aliases)
}
\arguments{
\item{meta}{A list containing parsed EGA metadata, where elements correspond
to individual sheets.}

\item{aliases}{A list of aliases, reference values used for validation (not
used in this function but included for consistency with other validators).}
}
\value{
A validation object containing the results of the check.
}
\description{
Validates the metadata by checking the following:
\itemize{
\item study titles are unique
\item study descriptions are unique
}
}
\examples{
meta <- list(studies = data.frame(
    title = c("StudyA", "StudyA", "StudyB"),
    description = c("Desc1", "Desc2", NA)
))
Rega:::.studies_extra_validator(meta, list())

}
\keyword{internal}
