% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{ega_oauth}
\alias{ega_oauth}
\title{Set The OAUTH With EGA Username And Password}
\usage{
ega_oauth(
  req,
  username = .get_ega_username(),
  password = .get_ega_password(),
  token_url = NULL
)
}
\arguments{
\item{req}{A httr2 request.}

\item{username}{Character. EGA User name. Defaults to the value returned by
\code{.get_ega_username()}.}

\item{password}{Character. EGA user Password. Defaults to the value returned
by \code{.get_ega_password()}.}

\item{token_url}{Character. The URL for the EGA token endpoint. Defaults to
the standard EGA token URL if not provided.}
}
\value{
returns a modified HTTP request that will use OAuth
}
\description{
\code{ega_oauth} implements the EGA OAuth resource owner password flow, as defined
by Section 4.3 of RFC 6749. It allows the user to supply their password once,
exchanging it for an access token that can be cached locally. Please avoid
entering the password directly when calling this function as it will be
captured by \code{.Rhistory}.
}
\examples{
req <- httr2::request("https://example.com/")

# Request OAuth with default credentials
oauth_req <- ega_oauth(req)

# Request OAuth with custom credentials
oauth_req <- ega_oauth(req, username = "user", password = "pass")

}
