% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_helpers.R
\name{fetch_files}
\alias{fetch_files}
\title{Fetch file information from EGA inbox}
\usage{
fetch_files(file_list, client = NULL)
}
\arguments{
\item{file_list}{A character vector or list of file prefixes to check.}

\item{client}{List of functions. EGA API client created by \code{create_client}
function from EGA API schema. If \code{NULL}, default client will be
created by \code{create_client(extract_api())}. Defaults to \code{NULL}.}
}
\value{
Data frame. Parsed response from client API for requested files.
}
\description{
Query the remote client for requested file prefix, test whether a file is
found for every element of \code{file_list} and return the server response.
}
\examples{
mock_client <- list(
    get__files = function(prefix = NULL) {
        data.frame(provisional_id = 12345, ega_relative_path = prefix)
      }
)
fetch_files(c("file_a", "file_b"), mock_client)

}
