% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{file_formatter}
\alias{file_formatter}
\title{Format File Table with EGA File Paths}
\usage{
file_formatter(tab, params)
}
\arguments{
\item{tab}{Data frame. The input table containing file information. Columns
\code{file}, \code{ega_inbox_relative_path} need to be present in the data.}

\item{params}{List. Additional parameters for formatting. Takes a formatter
params value from parser parameter yaml file. Includes \code{crypt_ext} for
encryption file extensions and \code{prepend_slash} to control path prefix.}
}
\value{
A formatted data frame with cleaned column names, and updated
\code{ega_file} paths based on file and relative path information.
}
\description{
Format File Table with EGA File Paths
}
\examples{
\dontrun{
# Format a file table
tab <- data.frame(
  file = c("file1.txt", "file2.txt"),
  ega_inbox_relative_path = c("/path1", "/path2")
)
params <- list(crypt_ext = "enc", prepend_slash = TRUE)
formatted_tab <- file_formatter(tab, params)
}

}
