% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{fold_column}
\alias{fold_column}
\title{Fold Columns with a Common Prefix into a Single Column Nested as List}
\usage{
fold_column(tab, column_prefix, new_name)
}
\arguments{
\item{tab}{Data frame. The input table with columns to fold.}

\item{column_prefix}{Character. The prefix of columns to nest into a single
column represented as list.}

\item{new_name}{Character. The name of the new folded column.}
}
\value{
A data frame with the specified columns nested into a single column.
}
\description{
If \code{NA} values are present in any of the columns to be nested, they
will be removed. If the column is not present it will be added with
\code{NA} as a single value.
}
\examples{
tab <- data.frame(id = c(1, 2), name.1 = c("A1", NA), name.2 = c("B1", "B2"))
fold_column(tab, "name", "folded_column")

}
