% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_chromosomes}
\alias{format_chromosomes}
\title{Format Chromosome Metadata}
\usage{
format_chromosomes(metadata)
}
\arguments{
\item{metadata}{List. A list of data frames representing metadata sheets,
containing \code{analyses}. Each row in \code{analyses} can have entry in
\code{chromosomes} or \code{chromosome_groups} column.}
}
\value{
A list of formatted chromosome data extracted or computed from the
input metadata.
}
\description{
Formats and processes chromosome-related metadata from an input object by
applying chromosome group lookups or splitting chromosome strings from the
EGA enums.
}
\examples{
# Mock metadata data frame
metadata <- list(
    analyses = data.frame(
        chromosomes = I(list(
            NA,
            list("group1--1--chr1--name1", "group2--3--chr3--name3")
        )),
        chromosome_groups = c("group1", NA),
        stringsAsFactors = FALSE
    ),
    select_input_data = list(
        chromosomes = c("group1--1--chr1--name1", "group1--2--chr2--name2")
    )
)

format_chromosomes(metadata)

}
