% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_enum}
\alias{get_enum}
\title{Retrieve Enum Values from an API}
\usage{
get_enum(client, enum_name, enum_prefix = "get__enums_")
}
\arguments{
\item{client}{List. The API client, typically generated by \code{create_client},
containing functions for API operations.}

\item{enum_name}{Character. The name of the enum to retrieve.}

\item{enum_prefix}{Character. Optional. The prefix used in the client for
enum functions. Defaults to \code{"get__enums_"}.}
}
\value{
The values of the specified enum.
}
\description{
This function retrieves the values of a specified enum from an API by
invoking the corresponding client function.
}
\examples{
\dontrun{
# Retrieve enum values from the client
enums <- get_enum(client, enum_name = "status")
}

}
