% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_formatter}
\alias{get_formatter}
\title{Retrieve a Formatter Function by Type of Submission Metadata Table}
\usage{
get_formatter(x, params)
}
\arguments{
\item{x}{Character. The name of the submission metadata table/sheet.}

\item{params}{List. A list containing a \code{formatter} element from parser
parameter yaml file.}
}
\value{
The formatter function corresponding to the specified table.
}
\description{
Retrieve a Formatter Function by Type of Submission Metadata Table
}
\examples{
# Load formatter params
params <- yaml::read_yaml(system.file(
    "extdata/default_parser_params.yaml",
    package = "Rega"
))

# Dummy data, first row will be moved to column names
tab <- data.frame(
    x1 = c("file", "value1", "value2"),
    x2 = c("ega_inbox_relative_path", NA, "proj1")
)

ff <- get_formatter("files", params)
ff_params <- get_formatter_params("files", params)

ff(tab, ff_params)

}
