% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_properties}
\alias{get_properties}
\title{Extract and Format Properties from a Schema}
\usage{
get_properties(schema, filter_ids = TRUE)
}
\arguments{
\item{schema}{List. The schema containing \code{properties} and \code{required} fields.}

\item{filter_ids}{Logical. Whether to filter out ID fields from the
properties. Defaults to \code{TRUE}.}
}
\value{
A character vector of formatted property names and indications of
required fields.
}
\description{
This function extracts property names from a schema, optionally filters out
ID fields, and applies formatting such as marking required fields and
prettifying the labels.
}
\examples{
schemas <- get_schemas(extract_api())

# Extract and format properties from a schema
get_properties(schemas[[5]])

# Extract properties without filtering ID fields
get_properties(schemas[[6]], filter_ids = FALSE)

}
