% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sentence_number}
\alias{get_sentence_number}
\title{Count the number of sentences in text}
\usage{
get_sentence_number(text)
}
\arguments{
\item{text}{A length-one character vector.}
}
\value{
An integer scalar giving the number of sentences in \code{text}.
}
\description{
Count how many sentences are in a character string, based on terminal
punctuation marks \code{.}, \code{!}, or \code{?} following an
alphanumeric character.
}
\examples{
get_sentence_number("First sentence. Second sentence? Third!")

}
