% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{get_submission}
\alias{get_submission}
\title{Retrieve Submission Data and Log Responses}
\usage{
get_submission(id, client, logfile = NULL, ...)
}
\arguments{
\item{id}{A string representing the submission identifier. Can be either an
accession or provisional ID.}

\item{client}{An API client object with \code{get} methods.}

\item{logfile}{A string specifying the path to a log file. If \code{NULL},
no log is written.}

\item{...}{Additional arguments for future extensions (currently unused).}
}
\value{
A list of responses including submission data and associated
datasets, analyses, runs, experiments, samples, and studies.
}
\description{
Retrieves data associated with a submission ID using the client and logs the
responses if a logfile is specified.
}
\examples{
mock_client <- list(
  "get__submissions__accession_id" = function(id) list(data = id),
  "get__submissions__accession_id__datasets" = function(id) list(datasets = id)
)
get_submission("EGAB12345678901", mock_client)

}
