% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_word_number}
\alias{get_word_number}
\title{Get number of words in the text}
\usage{
get_word_number(text)
}
\arguments{
\item{text}{}
}
\value{
Integer, number of words in the text
}
\description{
Get number of words in the text
}
