% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_helpers.R
\name{is_accession}
\alias{is_accession}
\title{Check if a String is a Valid Accession Identifier.}
\usage{
is_accession(x, schema = NULL)
}
\arguments{
\item{x}{A character vector to be tested for validity as accessions.}

\item{schema}{A character string specifying the schema. Valid options include
"study", "studies", "sample", "samples", "experiment", "experiments",
"analysis", "analyses", "run", "runs", "policy", "DAC", "dataset",
"datasets", "submission" and \code{NULL}. \code{NULL} will check against
any of the schemas. Defaults to \code{NULL}.}
}
\value{
A logical vector indicating which values are accession IDs.
}
\description{
Verifies whether the input string matches the format of a valid accession
identifier based on a specified schema.
}
\examples{
is_accession("EGAB00000000001", "submission") # TRUE
is_accession("EGA12345678901", "sample") # FALSE

}
