% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_helpers.R
\name{is_accession}
\alias{is_accession}
\title{Check if a String is a Valid Accession Identifier}
\usage{
is_accession(x, schema = "submission")
}
\arguments{
\item{x}{A character vector to be tested for validity as an accession.}

\item{schema}{A character string specifying the schema. Valid options
include "study", "studies", "sample", "samples", "experiment",
"experiments", "analysis", "analyses", "run", "runs", "policy",
"DAC", "dataset", "datasets", and "submission". Defaults to "submission".}
}
\value{
A logical vector indicating whether each element of \code{x} is a
valid accession identifier for the specified schema.
}
\description{
This function verifies whether the input string matches the format of a
valid accession identifier based on a specified schema.
}
\examples{
is_accession("EGAB00000000001", "submission") # TRUE
is_accession("EGA12345678901", "sample") # FALSE

}
