% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{parse_ega_body}
\alias{parse_ega_body}
\title{Parse The Information From EGA httr2 Response Object.}
\usage{
parse_ega_body(resp)
}
\arguments{
\item{resp}{An HTTP response object from the EGA API.}
}
\value{
A tibble containing the parsed and formatted response data. If the
response is plain text without a JSON-like structure, a one-column tibble
is returned with the raw content.
}
\description{
Parses the body of a body of \code{httr2} response object from the EGA API,
handling JSON and plain text content, and formats it into a tibble for
further processing.
}
\examples{
\dontrun{
url <- "https://submission.ega-archive.org/api/files"
req <- req_method(request(url), "GET")
req <- Rega:::ega_oauth(req)
req <- req_headers(req, `Content-Type` = "application/json")
req <- req_url_query(req, status = NULL, prefix = NULL)
resp <- req_perform(req, verbosity = 3)
Rega:::ega_parse_body(resp)
}

}
\keyword{internal}
