% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiclient.R
\name{parse_text_body}
\alias{parse_text_body}
\title{Parse Plain Text or JSON-like Response Body}
\usage{
parse_text_body(resp)
}
\arguments{
\item{resp}{An \code{httr2_response} object with "text/plain" content.}

\item{resource_name}{String used as the column name for raw text output.}
}
\value{
A list of parsed data or a tibble if the content is raw text.
}
\description{
Processes a text response by either parsing it as JSON (if structured with
curly braces or square brackets) or returning it as a list. Null JSON
elements are converted to empty lists to facilitate unnesting.
}
\examples{
text_resp <- httr2::response(
    method = "POST",
    url = "https://www.example.com/api/submissions",
    status = 200,
    headers = list("content-type" = "text/plain"),
    body = charToRaw("Sample response text")
)
parse_text_body(text_resp, "files")

}
