% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{rollback_submission}
\alias{rollback_submission}
\title{Rollback Submission Endpoints and Log Responses}
\usage{
rollback_submission(id, endpoints, client = NULL, logfile = NULL, ...)
}
\arguments{
\item{id}{A string representing the submission identifier. Must be an
accession ID.}

\item{endpoints}{A character vector of endpoint names to rollback.}

\item{client}{List of functions. EGA API client created by \code{create_client}
function from EGA API schema with \code{put} methods and rollback
operations. If \code{NULL}, default client will be created by
\code{create_client(extract_api())}. Defaults to \code{NULL}.}

\item{logfile}{A string specifying the path to a log file. If \code{NULL}, no
log is written. Defaults to \code{NULL}.}

\item{...}{Additional arguments for future extensions (currently unused).}
}
\value{
A list of responses from the rollback operations for each endpoint.
}
\description{
Rolls back specified endpoints for a submission identified by its accession
ID using the client and logs the responses if a logfile is specified.
}
\examples{
mock_client <- list(
    "put__submissions__accession_id__datasets_rollback" =
        function(id) list(status = "rolled back")
)
rollback_submission("provisional123", mock_client, c("datasets"))

}
