% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{row_table_formatter}
\alias{row_table_formatter}
\title{Format a Row Table}
\usage{
row_table_formatter(tab, params)
}
\arguments{
\item{tab}{Data frame. The input table from a sumbission metadata file}

\item{params}{List. Additional parameters for formatting. Takes a formatter
params value from parser parameter yaml file.}
}
\value{
A cleaned and formatted tibble with correctly organized rows and
columns, whitespace trimmed, and folding applied to specified
columns.
}
\description{
Format a Row Table
}
\examples{
\dontrun{
# Format a row table
tab <- data.frame(A = c("Name1", NA, "Name2"), B = c(1, 2, NA), C = c(NA, 3, 4))
params <- list(fold = c("A"))
formatted_tab <- row_table_formatter(tab, params)
}

}
