% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{use_submission}
\alias{use_submission}
\title{Retrieve or Delete Submission Data}
\usage{
use_submission(id, method, client = NULL)
}
\arguments{
\item{id}{Character or numeric. Represents the submission identifier. Can be
either an accession or provisional ID.}

\item{method}{A string specifying the operation to perform. Valid options are
"get" or "delete".}

\item{client}{List of functions. EGA API client created by \code{create_client}
function from EGA API schema with \code{get} and \code{delete} methods.
If \code{NULL}, default client will be created by
\code{create_client(extract_api())}. Defaults to \code{NULL}.}
}
\value{
A named list containing responses for datasets, analyses, runs,
experiments, samples, and studies.
}
\description{
Handles retrieval or deletion of data associated with a submission
accession/provisional ID using a specified client and method.
}
\examples{
mock_client <- list(
    "get__submissions__accession_id__datasets" = function(id) {
        list(data = id)
    },
    "delete__submissions__provisional_id__datasets" =
        function(id) list(status = "deleted")
)
use_submission("EGAB12345678901", mock_client, "get")

}
