% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_schema}
\alias{validate_schema}
\title{Validate a Payload Against a JSON Schema}
\usage{
validate_schema(payload, schema)
}
\arguments{
\item{payload}{List or JSON string. The payload to validate against the
schema.}

\item{schema}{List. The JSON schema defining the validation rules.}
}
\value{
Logical value indicating whether the payload is valid. If invalid,
the result includes an \code{errors} attribute detailing the validation errors.
}
\description{
Function handles \code{oneOf} directives in a way that it in a case of validation
fail, it displays the overall result of the validation as first and then it
tests separately against all \code{oneOf} sub schemas.
}
\examples{
\dontrun{
# Validate a payload against a schema
schema <- list(type = "object", properties = list(name = list(type = "string")))
payload <- list(name = "John")
is_valid <- validate_schema(payload, schema)
}

}
