% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validation_to_msg}
\alias{validation_to_msg}
\title{Convert Validation Results to a Message}
\usage{
validation_to_msg(v)
}
\arguments{
\item{v}{Logical. The validation result, which may include an
\code{errors} attribute detailing validation errors.}
}
\value{
A character string summarizing the validation results. If validation
errors are present, they are included in the message; otherwise, a success
message is returned.
}
\description{
Convert Validation Results to a Message
}
\examples{
validation_result <- FALSE
attr(validation_result, "errors") <- data.frame(
    field = c("name"),
    message = c("Missing")
)
msg <- validation_to_msg(validation_result)
message(msg)

}
